Change all execution tests to use the universe
