+++ massa-execution-worker/Cargo.toml
@@ -15,9 +15,6 @@ gas_calibration = [
     "massa_execution_exports/gas_calibration",
     "massa_final_state/test-exports",
     "massa_pos_worker",
-    "massa_ledger_worker",
-    "massa_db_worker",
-    "tempfile",
 ]
 test-exports = [
     "massa_execution_exports/test-exports",
@@ -26,20 +23,14 @@ test-exports = [
     "massa_final_state/test-exports",
     "massa-sc-runtime/testing",
     "massa_wallet/test-exports",
-    "tempfile",
     "massa_pos_worker",
-    "massa_ledger_worker",
     "massa_metrics/test-exports",
-    "massa_db_worker",
     "massa_metrics/test-exports",
 ]
 benchmarking = [
     "massa-sc-runtime/gas_calibration",
     "criterion",
     "massa_pos_worker",
-    "massa_ledger_worker",
-    "massa_db_worker",
-    "tempfile",
 ]
 metrics = []
 
@@ -59,8 +50,6 @@ sha2 = { workspace = true }
 sha3 = { workspace = true }
 libsecp256k1 = { workspace = true }
 criterion = { workspace = true, "optional" = true }
-tempfile = { workspace = true, "optional" = true } # BOM UPGRADE     Revert to {"version": "3.3", "optional": true} if problem
-massa_ledger_worker = { workspace = true, "optional" = true }
 massa_pos_worker = { workspace = true, "optional" = true }
 massa_async_pool = { workspace = true }
 massa_channel = { workspace = true }
@@ -79,7 +68,6 @@ massa_pos_exports = { workspace = true }
 massa_final_state = { workspace = true }
 massa_versioning = { workspace = true }
 massa_db_exports = { workspace = true }
-massa_db_worker = { workspace = true, "optional" = true }
 massa_wallet = { workspace = true }
 massa-proto-rs = { workspace = true }
 
@@ -94,9 +82,6 @@ massa-sc-runtime = { workspace = true, features = ["testing"] }
 massa_wallet = { workspace = true, features = ["test-exports"] }
 massa_metrics = { workspace = true, features = ["test-exports"] }
 massa_test_framework = {workspace = true}
-massa_db_worker = { workspace = true }
 tokio = { workspace = true, features = ["sync"] }
-massa_ledger_worker = { workspace = true }
 serial_test = { workspace = true }
-tempfile = { workspace = true }                                          # BOM UPGRADE     Revert to "3.2" if problem
 hex-literal = { workspace = true }
