+++ massa-execution-worker/src/interface_impl.rs
@@ -88,14 +88,25 @@ impl InterfaceImpl {
         sender_addr: Address,
         operation_datastore: Option<Datastore>,
     ) -> InterfaceImpl {
+        use massa_final_state::test_exports::get_sample_state;
         use massa_ledger_exports::{LedgerEntry, SetUpdateOrDelete};
         use massa_models::config::MIP_STORE_STATS_BLOCK_CONSIDERED;
         use massa_module_cache::{config::ModuleCacheConfig, controller::ModuleCache};
+        use massa_pos_exports::SelectorConfig;
+        use massa_pos_worker::start_selector_worker;
         use massa_versioning::versioning::{MipStatsConfig, MipStore};
         use parking_lot::RwLock;
 
         let config = ExecutionConfig::default();
-        let (final_state, _tempfile, _tempdir) = super::tests::get_sample_state(0).unwrap();
+        let mip_stats_config = MipStatsConfig {
+            block_count_considered: MIP_STORE_STATS_BLOCK_CONSIDERED,
+            warn_announced_version_ratio: Ratio::new_raw(30, 100),
+        };
+        let mip_store = MipStore::try_from(([], mip_stats_config)).unwrap();
+        let (_, selector_controller) = start_selector_worker(SelectorConfig::default())
+            .expect("could not start selector controller");
+        let (final_state, _tempfile, _tempdir) =
+            get_sample_state(config.last_start_period, selector_controller, mip_store).unwrap();
         let module_cache = Arc::new(RwLock::new(ModuleCache::new(ModuleCacheConfig {
             hd_cache_path: config.hd_cache_path.clone(),
             gas_costs: config.gas_costs.clone(),
