+++ massa-execution-worker/src/tests/interface.rs
@@ -1,22 +1,21 @@
 // Copyright (c) 2022 MASSA LABS <info@massa.net>
 
-#[cfg(test)]
-mod tests {
-    use hex_literal::hex;
-    use massa_models::address::Address;
-    use massa_sc_runtime::Interface;
-    use std::str::FromStr;
+use hex_literal::hex;
+use massa_models::address::Address;
+use massa_sc_runtime::Interface;
+use serial_test::serial;
+use std::str::FromStr;
 
-    use crate::interface_impl::InterfaceImpl;
-    #[test]
-    fn test_hash_sha256() {
-        let interface = InterfaceImpl::new_default(
-            Address::from_str("AU12cMW9zRKFDS43Z2W88VCmdQFxmHjAo54XvuVV34UzJeXRLXW9M").unwrap(),
-            None,
-        );
-        let actual_hash = interface.hash_sha256(b"something").unwrap();
-        let expected_hash =
-            &hex!("3fc9b689459d738f8c88a3a48aa9e33542016b7a4052e001aaa536fca74813cb")[..];
-        assert_eq!(actual_hash, expected_hash);
-    }
+use crate::interface_impl::InterfaceImpl;
+#[test]
+#[serial]
+fn test_hash_sha256() {
+    let interface = InterfaceImpl::new_default(
+        Address::from_str("AU12cMW9zRKFDS43Z2W88VCmdQFxmHjAo54XvuVV34UzJeXRLXW9M").unwrap(),
+        None,
+    );
+    let actual_hash = interface.hash_sha256(b"something").unwrap();
+    let expected_hash =
+        &hex!("3fc9b689459d738f8c88a3a48aa9e33542016b7a4052e001aaa536fca74813cb")[..];
+    assert_eq!(actual_hash, expected_hash);
 }
