+++ massa-execution-worker/src/tests/mod.rs
@@ -1,14 +1,5 @@
 // Copyright (c) 2022 MASSA LABS <info@massa.net>
 
-#[cfg(any(
-    test,
-    feature = "gas_calibration",
-    feature = "benchmarking",
-    feature = "test-exports",
-    test
-))]
-mod mock;
-
 #[cfg(test)]
 mod scenarios_mandatories;
 
@@ -19,12 +10,3 @@ mod universe;
 mod tests_active_history;
 
 mod interface;
-
-
-#[cfg(any(
-    feature = "gas_calibration",
-    feature = "benchmarking",
-    feature = "test-exports",
-    test
-))]
-pub use mock::get_sample_state;
