+++ massa-execution-worker/src/tests/scenarios_mandatories.rs
@@ -1,49 +1,30 @@
 // Copyright (c) 2022 MASSA LABS <info@massa.net>
 
-use crate::active_history::ActiveHistory;
-use crate::speculative_async_pool::SpeculativeAsyncPool;
-use crate::start_execution_worker;
-use crate::tests::mock::{create_block, get_random_address_full, get_sample_state};
-use massa_async_pool::AsyncMessage;
 use massa_db_exports::DBBatch;
 use massa_execution_exports::{
-    ExecutionBlockMetadata, ExecutionChannels, ExecutionConfig, ExecutionController,
-    ExecutionError, ReadOnlyExecutionRequest, ReadOnlyExecutionTarget,
+    ExecutionBlockMetadata, ExecutionConfig, ReadOnlyExecutionRequest, ReadOnlyExecutionTarget,
 };
-use massa_hash::Hash;
-use massa_metrics::MassaMetrics;
-use massa_models::config::{
-    LEDGER_ENTRY_BASE_COST, LEDGER_ENTRY_DATASTORE_BASE_SIZE, MIP_STORE_STATS_BLOCK_CONSIDERED,
-};
+use massa_models::config::{LEDGER_ENTRY_BASE_COST, LEDGER_ENTRY_DATASTORE_BASE_SIZE};
 use massa_models::prehash::PreHashMap;
 use massa_models::test_exports::gen_endorsements_for_denunciation;
-use massa_models::{
-    address::{Address, UserAddress, UserAddressV0},
-    amount::Amount,
-    slot::Slot,
-};
+use massa_models::{address::Address, amount::Amount, slot::Slot};
 use massa_models::{
     block_id::BlockId,
-    datastore::Datastore,
     denunciation::Denunciation,
     execution::EventFilter,
-    operation::{Operation, OperationSerializer, OperationType, SecureShareOperation},
+    operation::{Operation, OperationSerializer, OperationType},
     secure_share::SecureShareContent,
 };
+use massa_pos_exports::MockSelectorControllerWrapper;
 use massa_signature::KeyPair;
 use massa_storage::Storage;
 use massa_test_framework::TestUniverse;
 use massa_time::MassaTime;
-use massa_versioning::versioning::{MipStatsConfig, MipStore};
-use massa_wallet::test_exports::create_test_wallet;
 use num::rational::Ratio;
-use parking_lot::RwLock;
 use serial_test::serial;
-use std::sync::Arc;
 use std::{
     cmp::Reverse, collections::BTreeMap, collections::HashMap, str::FromStr, time::Duration,
 };
-use tokio::sync::broadcast;
 
 use super::universe::{ExecutionForeignControllers, ExecutionTestUniverse};
 
@@ -51,41 +32,51 @@ const TEST_SK_1: &str = "S18r2i8oJJyhF7Kprx98zwxAc3W4szf7RKuVMX6JydZz8zSxHeC";
 const TEST_SK_2: &str = "S1FpYC4ugG9ivZZbLVrTwWtF9diSRiAwwrVX5Gx1ANSRLfouUjq";
 const TEST_SK_3: &str = "S1LgXhWLEgAgCX3nm6y8PVPzpybmsYpi6yg6ZySwu5Z4ERnD7Bu";
 
+fn selector_boilerplate(
+    mock_selector: &mut MockSelectorControllerWrapper,
+    block_creator: &KeyPair,
+) {
+    let block_creator_address = Address::from_public_key(&block_creator.get_public_key());
+    mock_selector.set_expectations(|selector_controller| {
+        selector_controller
+            .expect_feed_cycle()
+            .returning(move |_, _, _| Ok(()));
+        selector_controller
+            .expect_wait_for_draws()
+            .returning(move |cycle| Ok(cycle + 1));
+        selector_controller
+            .expect_get_producer()
+            .returning(move |_| Ok(block_creator_address));
+    });
+}
+
 #[test]
 #[serial]
 fn test_execution_shutdown() {
+    let block_producer = KeyPair::generate(0).unwrap();
     let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
-    foreign_controllers.selector_controller.set_expectations(|selector_controller| {
-        selector_controller.expect_feed_cycle().times(2).returning(move |_, _, _| {
-            Ok(())
-        });
-        selector_controller.expect_wait_for_draws().times(1).returning(move |cycle| {
-            Ok(cycle + 1)
-        });
-    });
-    let mut universe = ExecutionTestUniverse::new(foreign_controllers, ExecutionConfig::default());
-    universe.stop();
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
+    );
+    ExecutionTestUniverse::new(foreign_controllers, ExecutionConfig::default());
 }
 
 #[test]
 #[serial]
 fn test_sending_command() {
+    let block_producer = KeyPair::generate(0).unwrap();
     let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
-    foreign_controllers.selector_controller.set_expectations(|selector_controller| {
-        selector_controller.expect_feed_cycle().times(2).returning(move |_, _, _| {
-            Ok(())
-        });
-        selector_controller.expect_wait_for_draws().times(1).returning(move |cycle| {
-            Ok(cycle + 1)
-        });
-    });
-    let mut universe = ExecutionTestUniverse::new(foreign_controllers, ExecutionConfig::default());
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
+    );
+    let universe = ExecutionTestUniverse::new(foreign_controllers, ExecutionConfig::default());
     universe.module_controller.update_blockclique_status(
         Default::default(),
         Default::default(),
         Default::default(),
     );
-    universe.stop();
 }
 
 #[test]
@@ -97,18 +88,16 @@ fn test_readonly_execution() {
         cursor_delay: MassaTime::from_millis(0),
         ..ExecutionConfig::default()
     };
+    let block_producer = KeyPair::generate(0).unwrap();
     let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
-    foreign_controllers.selector_controller.set_expectations(|selector_controller| {
-        selector_controller.expect_feed_cycle().times(2).returning(move |_, _, _| {
-            Ok(())
-        });
-        selector_controller.expect_wait_for_draws().times(1).returning(move |cycle| {
-            Ok(cycle + 1)
-        });
-    });
-    let mut universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg);
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
+    );
+    let universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg);
 
-    let mut res = universe.module_controller
+    let mut res = universe
+        .module_controller
         .execute_readonly_request(ReadOnlyExecutionRequest {
             max_gas: 1_000_000,
             call_stack: vec![],
@@ -124,45 +113,6 @@ fn test_readonly_execution() {
     assert_eq!(res.out.slot, Slot::new(1, 0));
     assert!(res.gas_cost > 0);
     assert_eq!(res.out.events.take().len(), 1, "wrong number of events");
-
-    universe.stop();
-}
-
-/// generate a random address
-fn get_random_address() -> Address {
-    let kp = KeyPair::generate(0).unwrap();
-    Address::from_public_key(&kp.get_public_key())
-}
-
-/// Feeds the execution worker with genesis blocks to start it
-fn init_execution_worker(
-    config: &ExecutionConfig,
-    storage: &Storage,
-    execution_controller: Box<dyn ExecutionController>,
-) {
-    let genesis_keypair = KeyPair::generate(0).unwrap();
-    let genesis_addr = Address::from_public_key(&genesis_keypair.get_public_key());
-    let mut finalized_blocks: HashMap<Slot, BlockId> = HashMap::new();
-    let mut block_metadata: PreHashMap<BlockId, ExecutionBlockMetadata> = PreHashMap::default();
-    for thread in 0..config.thread_count {
-        let slot = Slot::new(0, thread);
-        let final_block = create_block(genesis_keypair.clone(), vec![], vec![], slot).unwrap();
-        finalized_blocks.insert(slot, final_block.id);
-        let mut final_block_storage = storage.clone_without_refs();
-        final_block_storage.store_block(final_block.clone());
-        block_metadata.insert(
-            final_block.id,
-            ExecutionBlockMetadata {
-                same_thread_parent_creator: Some(genesis_addr),
-                storage: Some(final_block_storage),
-            },
-        );
-    }
-    execution_controller.update_blockclique_status(
-        finalized_blocks,
-        Some(Default::default()),
-        block_metadata,
-    );
 }
 
 /// Test the gas usage in nested calls using call SC operation
@@ -183,19 +133,11 @@ fn test_nested_call_gas_usage() {
         ..ExecutionConfig::default()
     };
     let block_producer = KeyPair::generate(0).unwrap();
-    let block_producer_address = Address::from_public_key(&block_producer.get_public_key());
     let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
-    foreign_controllers.selector_controller.set_expectations(|selector_controller| {
-        selector_controller.expect_feed_cycle().times(2).returning(move |_, _, _| {
-            Ok(())
-        });
-        selector_controller.expect_wait_for_draws().times(1).returning(move |cycle| {
-            Ok(cycle + 1)
-        });
-        selector_controller.expect_get_producer().returning(move |_| {
-            Ok(block_producer_address)
-        });
-    });
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
+    );
     let mut universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg);
 
     // get random keypair
@@ -208,13 +150,15 @@ fn test_nested_call_gas_usage() {
     datastore.insert(b"smart-contract".to_vec(), datastore_bytecode);
 
     // create the block containing the smart contract execution operation
-    let operation = ExecutionTestUniverse::create_execute_sc_operation(&keypair, bytecode, datastore).unwrap();
+    let operation =
+        ExecutionTestUniverse::create_execute_sc_operation(&keypair, bytecode, datastore).unwrap();
     universe.storage.store_operations(vec![operation.clone()]);
     let block = ExecutionTestUniverse::create_block(
         &block_producer,
         Slot::new(1, 0),
-        Some(vec![operation]),
-        None,
+        vec![operation],
+        vec![],
+        vec![],
     );
     // store the block in storage
     universe.storage.store_block(block.clone());
@@ -226,7 +170,7 @@ fn test_nested_call_gas_usage() {
     block_metadata.insert(
         block.id,
         ExecutionBlockMetadata {
-            same_thread_parent_creator: Some(get_random_address()),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
             storage: Some(universe.storage.clone()),
         },
     );
@@ -236,11 +180,13 @@ fn test_nested_call_gas_usage() {
         block_metadata.clone(),
     );
     std::thread::sleep(Duration::from_millis(100));
-    let events = universe.module_controller.get_filtered_sc_output_event(EventFilter {
-        start: Some(Slot::new(0, 1)),
-        end: Some(Slot::new(20, 1)),
-        ..Default::default()
-    });
+    let events = universe
+        .module_controller
+        .get_filtered_sc_output_event(EventFilter {
+            start: Some(Slot::new(0, 1)),
+            end: Some(Slot::new(20, 1)),
+            ..Default::default()
+        });
     // match the events
     assert!(!events.is_empty(), "One event was expected");
     let address = events[0].clone().data;
@@ -261,9 +207,10 @@ fn test_nested_call_gas_usage() {
     let block = ExecutionTestUniverse::create_block(
         &block_producer,
         Slot::new(2, 0),
-        Some(vec![operation]),
-        None    
-    );   
+        vec![operation],
+        vec![],
+        vec![],
+    );
     // store the block in storage
     storage.store_block(block.clone());
     // set our block as a final block so the message is sent
@@ -273,7 +220,7 @@ fn test_nested_call_gas_usage() {
     block_metadata.insert(
         block.id,
         ExecutionBlockMetadata {
-            same_thread_parent_creator: Some(get_random_address()),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
             storage: Some(storage.clone()),
         },
     );
@@ -284,10 +231,12 @@ fn test_nested_call_gas_usage() {
     );
     std::thread::sleep(Duration::from_millis(100));
     // Get the events that give us the gas usage (refer to source in ts) without fetching the first slot because it emit a event with an address.
-    let events = universe.module_controller.get_filtered_sc_output_event(EventFilter {
-        start: Some(Slot::new(2, 0)),
-        ..Default::default()
-    });
+    let events = universe
+        .module_controller
+        .get_filtered_sc_output_event(EventFilter {
+            start: Some(Slot::new(2, 0)),
+            ..Default::default()
+        });
     assert!(!events.is_empty());
     // Check that we always subtract gas through the execution (even in sub calls)
     let events_formatted = events
@@ -300,8 +249,6 @@ fn test_nested_call_gas_usage() {
         events_formatted, sorted_events,
         "Gas is not going down through the execution."
     );
-    // stop the execution controller
-    universe.stop();
 }
 
 /// Test the ABI get call coins
@@ -318,21 +265,13 @@ fn test_get_call_coins() {
         ..ExecutionConfig::default()
     };
     let block_producer = KeyPair::generate(0).unwrap();
-    let block_producer_address = Address::from_public_key(&block_producer.get_public_key());
     let keypair = KeyPair::from_str(TEST_SK_1).unwrap();
 
     let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
-    foreign_controllers.selector_controller.set_expectations(|selector_controller| {
-        selector_controller.expect_feed_cycle().times(2).returning(move |_, _, _| {
-            Ok(())
-        });
-        selector_controller.expect_wait_for_draws().times(1).returning(move |cycle| {
-            Ok(cycle + 1)
-        });
-        selector_controller.expect_get_producer().returning(move |_| {
-            Ok(block_producer_address)
-        });
-    });
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
+    );
     let mut universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg);
 
     // load bytecodes
@@ -343,13 +282,15 @@ fn test_get_call_coins() {
     datastore.insert(b"smart-contract".to_vec(), datastore_bytecode);
 
     // create the block containing the smart contract execution operation
-    let operation = ExecutionTestUniverse::create_execute_sc_operation(&keypair, bytecode, datastore).unwrap();
+    let operation =
+        ExecutionTestUniverse::create_execute_sc_operation(&keypair, bytecode, datastore).unwrap();
     universe.storage.store_operations(vec![operation.clone()]);
     let block = ExecutionTestUniverse::create_block(
         &keypair,
         Slot::new(1, 0),
-        Some(vec![operation]),
-        None
+        vec![operation],
+        vec![],
+        vec![],
     );
     // store the block in storage
     universe.storage.store_block(block.clone());
@@ -362,7 +303,7 @@ fn test_get_call_coins() {
         block.id,
         ExecutionBlockMetadata {
             storage: Some(universe.storage.clone()),
-            same_thread_parent_creator: Some(get_random_address()),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
         },
     );
     universe.module_controller.update_blockclique_status(
@@ -375,11 +316,13 @@ fn test_get_call_coins() {
 
     // assert_eq!(balance, balance_expected);
     // retrieve events emitted by smart contracts
-    let events = universe.module_controller.get_filtered_sc_output_event(EventFilter {
-        start: Some(Slot::new(0, 1)),
-        end: Some(Slot::new(20, 1)),
-        ..Default::default()
-    });
+    let events = universe
+        .module_controller
+        .get_filtered_sc_output_event(EventFilter {
+            start: Some(Slot::new(0, 1)),
+            end: Some(Slot::new(20, 1)),
+            ..Default::default()
+        });
     // match the events
     assert!(!events.is_empty(), "One event was expected");
     let address = events[0].clone().data;
@@ -400,8 +343,9 @@ fn test_get_call_coins() {
     let block = ExecutionTestUniverse::create_block(
         &keypair,
         Slot::new(2, 0),
-        Some(vec![operation]),
-        None
+        vec![operation],
+        vec![],
+        vec![],
     );
     // store the block in storage
     universe.storage.store_block(block.clone());
@@ -412,7 +356,7 @@ fn test_get_call_coins() {
     block_metadata.insert(
         block.id,
         ExecutionBlockMetadata {
-            same_thread_parent_creator: Some(get_random_address()),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
             storage: Some(universe.storage.clone()),
         },
     );
@@ -423,18 +367,17 @@ fn test_get_call_coins() {
     );
     std::thread::sleep(Duration::from_millis(100));
     // Get the events that give us the gas usage (refer to source in ts) without fetching the first slot because it emit a event with an address.
-    let events = universe.module_controller.get_filtered_sc_output_event(EventFilter {
-        start: Some(Slot::new(2, 0)),
-        ..Default::default()
-    });
+    let events = universe
+        .module_controller
+        .get_filtered_sc_output_event(EventFilter {
+            start: Some(Slot::new(2, 0)),
+            ..Default::default()
+        });
     println!("events {:#?}", events);
     assert!(events[0].data.contains(&format!(
         "tokens sent to the SC during the call : {}",
         coins_sent.to_raw()
     )));
-
-    // stop the execution controller
-    universe.stop();
 }
 
 /// # Context
@@ -457,51 +400,21 @@ fn test_get_call_coins() {
 #[test]
 #[serial]
 fn send_and_receive_async_message() {
-    // setup the period duration and the maximum gas for asynchronous messages execution
     let exec_cfg = ExecutionConfig {
         t0: MassaTime::from_millis(100),
         cursor_delay: MassaTime::from_millis(0),
         max_async_gas: 100_000,
         ..ExecutionConfig::default()
     };
-    // get a sample final state
-    let (sample_state, _keep_file, _keep_dir) = get_sample_state(0).unwrap();
-
-    // init the MIP store
-    let mip_stats_config = MipStatsConfig {
-        block_count_considered: MIP_STORE_STATS_BLOCK_CONSIDERED,
-        warn_announced_version_ratio: Ratio::new_raw(30, 100),
-    };
-    let mip_store = MipStore::try_from(([], mip_stats_config)).unwrap();
-
-    let slot_execution_output_sender = broadcast::channel(5000).0;
-
-    let channels = ExecutionChannels {
-        slot_execution_output_sender,
-    };
-
-    // init the storage
-    let mut storage = Storage::create_root();
-    // start the execution worker
-    let (mut manager, controller) = start_execution_worker(
-        exec_cfg.clone(),
-        sample_state.clone(),
-        sample_state.read().pos_state.selector.clone(),
-        mip_store,
-        channels,
-        Arc::new(RwLock::new(create_test_wallet(Some(PreHashMap::default())))),
-        MassaMetrics::new(
-            false,
-            "0.0.0.0:9898".parse().unwrap(),
-            32,
-            std::time::Duration::from_secs(5),
-        )
-        .0,
-    );
-    // initialize the execution system with genesis blocks
-    init_execution_worker(&exec_cfg, &storage, controller.clone());
-    // keypair associated to thread 0
+    let block_producer = KeyPair::generate(0).unwrap();
     let keypair = KeyPair::from_str(TEST_SK_1).unwrap();
+
+    let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
+    );
+    let mut universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg);
 
     // load bytecodes
     // you can check the source code of the following wasm file in massa-unit-tests-src
@@ -511,17 +424,18 @@ fn send_and_receive_async_message() {
     datastore.insert(b"smart-contract".to_vec(), datastore_bytecode);
 
     // create the block contaning the smart contract execution operation
-    let operation = create_execute_sc_operation(&keypair, bytecode, datastore).unwrap();
-    storage.store_operations(vec![operation.clone()]);
-    let block = create_block(
-        KeyPair::generate(0).unwrap(),
+    let operation =
+        ExecutionTestUniverse::create_execute_sc_operation(&keypair, bytecode, datastore).unwrap();
+    universe.storage.store_operations(vec![operation.clone()]);
+    let block = ExecutionTestUniverse::create_block(
+        &block_producer,
+        Slot::new(1, 0),
         vec![operation],
         vec![],
-        Slot::new(1, 0),
-    )
-    .unwrap();
+        vec![],
+    );
     // store the block in storage
-    storage.store_block(block.clone());
+    universe.storage.store_block(block.clone());
 
     // set our block as a final block so the message is sent
     let mut finalized_blocks: HashMap<Slot, BlockId> = Default::default();
@@ -530,11 +444,11 @@ fn send_and_receive_async_message() {
     block_metadata.insert(
         block.id,
         ExecutionBlockMetadata {
-            same_thread_parent_creator: Some(get_random_address()),
-            storage: Some(storage),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
+            storage: Some(universe.storage.clone()),
         },
     );
-    controller.update_blockclique_status(
+    universe.module_controller.update_blockclique_status(
         finalized_blocks,
         Default::default(),
         block_metadata.clone(),
@@ -543,19 +457,19 @@ fn send_and_receive_async_message() {
     std::thread::sleep(Duration::from_millis(150));
 
     // retrieve events emitted by smart contracts
-    let events = controller.get_filtered_sc_output_event(EventFilter {
-        start: Some(Slot::new(1, 1)),
-        end: Some(Slot::new(20, 1)),
-        ..Default::default()
-    });
+    let events = universe
+        .module_controller
+        .get_filtered_sc_output_event(EventFilter {
+            start: Some(Slot::new(1, 1)),
+            end: Some(Slot::new(20, 1)),
+            ..Default::default()
+        });
 
     println!("events: {:?}", events);
 
     // match the events
     assert!(events.len() == 1, "One event was expected");
     assert_eq!(events[0].data, "message correctly received: 42,42,42,42");
-    // stop the execution controller
-    manager.stop();
 }
 
 /// # Context
@@ -579,52 +493,21 @@ fn send_and_receive_async_message() {
 #[test]
 #[serial]
 fn test_operation_execution_status() {
-    // setup the period duration and the maximum gas for asynchronous messages execution
+    // setup the period duration
     let exec_cfg = ExecutionConfig {
-        max_async_gas: 100_000,
         t0: MassaTime::from_millis(100),
         cursor_delay: MassaTime::from_millis(0),
         ..ExecutionConfig::default()
     };
-    // get a sample final state
-    let (sample_state, _keep_file, _keep_dir) = get_sample_state(0).unwrap();
-
-    // init the MIP store
-    let mip_stats_config = MipStatsConfig {
-        block_count_considered: MIP_STORE_STATS_BLOCK_CONSIDERED,
-        warn_announced_version_ratio: Ratio::new_raw(30, 100),
-    };
-    let mip_store = MipStore::try_from(([], mip_stats_config)).unwrap();
-
-    // init the storage
-    let mut storage = Storage::create_root();
-
-    let slot_execution_output_sender = broadcast::channel(5000).0;
-
-    let channels = ExecutionChannels {
-        slot_execution_output_sender,
-    };
-
-    // start the execution worker
-    let (mut manager, controller) = start_execution_worker(
-        exec_cfg.clone(),
-        sample_state.clone(),
-        sample_state.read().pos_state.selector.clone(),
-        mip_store,
-        channels,
-        Arc::new(RwLock::new(create_test_wallet(Some(PreHashMap::default())))),
-        MassaMetrics::new(
-            false,
-            "0.0.0.0:9898".parse().unwrap(),
-            32,
-            std::time::Duration::from_secs(5),
-        )
-        .0,
-    );
-    // initialize the execution system with genesis blocks
-    init_execution_worker(&exec_cfg, &storage, controller.clone());
-    // keypair associated to thread 0
+    let block_producer = KeyPair::generate(0).unwrap();
     let keypair = KeyPair::from_str(TEST_SK_1).unwrap();
+
+    let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
+    );
+    let mut universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg);
     // load bytecodes
     // you can check the source code of the following wasm file in massa-unit-tests-src
     let bytecode = include_bytes!("./wasm/send_message.wasm");
@@ -633,18 +516,19 @@ fn test_operation_execution_status() {
     datastore.insert(b"smart-contract".to_vec(), datastore_bytecode);
 
     // create the block contaning the smart contract execution operation
-    let operation = create_execute_sc_operation(&keypair, bytecode, datastore).unwrap();
+    let operation =
+        ExecutionTestUniverse::create_execute_sc_operation(&keypair, bytecode, datastore).unwrap();
     let tested_op_id = operation.id;
-    storage.store_operations(vec![operation.clone()]);
-    let block = create_block(
-        KeyPair::generate(0).unwrap(),
+    universe.storage.store_operations(vec![operation.clone()]);
+    let block = ExecutionTestUniverse::create_block(
+        &block_producer,
+        Slot::new(1, 0),
         vec![operation],
         vec![],
-        Slot::new(1, 0),
-    )
-    .unwrap();
+        vec![],
+    );
     // store the block in storage
-    storage.store_block(block.clone());
+    universe.storage.store_block(block.clone());
 
     // set our block as a final block so the message is sent
     let mut finalized_blocks: HashMap<Slot, BlockId> = Default::default();
@@ -653,11 +537,11 @@ fn test_operation_execution_status() {
     block_metadata.insert(
         block.id,
         ExecutionBlockMetadata {
-            same_thread_parent_creator: Some(get_random_address()),
-            storage: Some(storage),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
+            storage: Some(universe.storage.clone()),
         },
     );
-    controller.update_blockclique_status(
+    universe.module_controller.update_blockclique_status(
         finalized_blocks,
         Default::default(),
         block_metadata.clone(),
@@ -665,16 +549,15 @@ fn test_operation_execution_status() {
     // sleep for 150ms to reach the message execution period
     std::thread::sleep(Duration::from_millis(150));
 
-    let (op_candidate, op_final) = controller.get_ops_exec_status(&[tested_op_id])[0];
+    let (op_candidate, op_final) = universe
+        .module_controller
+        .get_ops_exec_status(&[tested_op_id])[0];
 
     // match the events
     assert!(
         op_candidate == Some(true) && op_final == Some(true),
         "Expected operation not found or not successfully executed"
     );
-
-    // stop the execution controller
-    manager.stop();
 }
 
 /// Context
@@ -691,51 +574,21 @@ fn test_operation_execution_status() {
 #[test]
 #[serial]
 fn local_execution() {
-    // setup the period duration and cursor delay
+    // setup the period duration
     let exec_cfg = ExecutionConfig {
         t0: MassaTime::from_millis(100),
         cursor_delay: MassaTime::from_millis(0),
         ..ExecutionConfig::default()
     };
-    // get a sample final state
-    let (sample_state, _keep_file, _keep_dir) = get_sample_state(0).unwrap();
-
-    // init the MIP store
-    let mip_stats_config = MipStatsConfig {
-        block_count_considered: MIP_STORE_STATS_BLOCK_CONSIDERED,
-        warn_announced_version_ratio: Ratio::new_raw(30, 100),
-    };
-    let mip_store = MipStore::try_from(([], mip_stats_config)).unwrap();
-
-    // init the storage
-    let mut storage = Storage::create_root();
-
-    let slot_execution_output_sender = broadcast::channel(5000).0;
-
-    let channels = ExecutionChannels {
-        slot_execution_output_sender,
-    };
-
-    // start the execution worker
-    let (mut manager, controller) = start_execution_worker(
-        exec_cfg.clone(),
-        sample_state.clone(),
-        sample_state.read().pos_state.selector.clone(),
-        mip_store,
-        channels,
-        Arc::new(RwLock::new(create_test_wallet(Some(PreHashMap::default())))),
-        MassaMetrics::new(
-            false,
-            "0.0.0.0:9898".parse().unwrap(),
-            32,
-            std::time::Duration::from_secs(5),
-        )
-        .0,
-    );
-    // initialize the execution system with genesis blocks
-    init_execution_worker(&exec_cfg, &storage, controller.clone());
-    // keypair associated to thread 0
+    let block_producer = KeyPair::generate(0).unwrap();
     let keypair = KeyPair::from_str(TEST_SK_1).unwrap();
+
+    let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
+    );
+    let mut universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg);
     // load bytecodes
     // you can check the source code of the following wasm files in massa-unit-tests-src
     let exec_bytecode = include_bytes!("./wasm/local_execution.wasm");
@@ -745,19 +598,27 @@ fn local_execution() {
     datastore.insert(b"smart-contract".to_vec(), datastore_bytecode);
 
     // create the block contaning the operations
-    let local_exec_op =
-        create_execute_sc_operation(&keypair, exec_bytecode, datastore.clone()).unwrap();
-    let local_call_op = create_execute_sc_operation(&keypair, call_bytecode, datastore).unwrap();
-    storage.store_operations(vec![local_exec_op.clone(), local_call_op.clone()]);
-    let block = create_block(
-        KeyPair::generate(0).unwrap(),
-        vec![local_exec_op.clone(), local_call_op.clone()],
-        vec![],
-        Slot::new(1, 0),
+    let local_exec_op = ExecutionTestUniverse::create_execute_sc_operation(
+        &keypair,
+        exec_bytecode,
+        datastore.clone(),
     )
     .unwrap();
+    let local_call_op =
+        ExecutionTestUniverse::create_execute_sc_operation(&keypair, call_bytecode, datastore)
+            .unwrap();
+    universe
+        .storage
+        .store_operations(vec![local_exec_op.clone(), local_call_op.clone()]);
+    let block = ExecutionTestUniverse::create_block(
+        &block_producer,
+        Slot::new(1, 0),
+        vec![local_exec_op, local_call_op],
+        vec![],
+        vec![],
+    );
     // store the block in storage
-    storage.store_block(block.clone());
+    universe.storage.store_block(block.clone());
 
     // set our block as a final block so the message is sent
     let mut finalized_blocks: HashMap<Slot, BlockId> = Default::default();
@@ -766,11 +627,11 @@ fn local_execution() {
     block_metadata.insert(
         block.id,
         ExecutionBlockMetadata {
-            same_thread_parent_creator: Some(get_random_address()),
-            storage: Some(storage),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
+            storage: Some(universe.storage.clone()),
         },
     );
-    controller.update_blockclique_status(
+    universe.module_controller.update_blockclique_status(
         finalized_blocks,
         Default::default(),
         block_metadata.clone(),
@@ -779,9 +640,11 @@ fn local_execution() {
     std::thread::sleep(Duration::from_millis(100));
 
     // retrieve events emitted by smart contracts
-    let events = controller.get_filtered_sc_output_event(EventFilter {
-        ..Default::default()
-    });
+    let events = universe
+        .module_controller
+        .get_filtered_sc_output_event(EventFilter {
+            ..Default::default()
+        });
 
     // match the events, check balance and call stack to make sure the executions were local
     assert!(events.len() == 8, "8 events were expected");
@@ -803,9 +666,6 @@ fn local_execution() {
         &Address::from_public_key(&keypair.get_public_key())
     );
     assert_eq!(events[6].data, "one local call completed");
-
-    // stop the execution controller
-    manager.stop();
 }
 
 /// Context
@@ -821,51 +681,21 @@ fn local_execution() {
 #[test]
 #[serial]
 fn sc_deployment() {
-    // setup the period duration and cursor delay
+    // setup the period duration
     let exec_cfg = ExecutionConfig {
         t0: MassaTime::from_millis(100),
         cursor_delay: MassaTime::from_millis(0),
         ..ExecutionConfig::default()
     };
-    // get a sample final state
-    let (sample_state, _keep_file, _keep_dir) = get_sample_state(0).unwrap();
-
-    // init the MIP store
-    let mip_stats_config = MipStatsConfig {
-        block_count_considered: MIP_STORE_STATS_BLOCK_CONSIDERED,
-        warn_announced_version_ratio: Ratio::new_raw(30, 100),
-    };
-    let mip_store = MipStore::try_from(([], mip_stats_config)).unwrap();
-
-    // init the storage
-    let mut storage = Storage::create_root();
-
-    let slot_execution_output_sender = broadcast::channel(5000).0;
-
-    let channels = ExecutionChannels {
-        slot_execution_output_sender,
-    };
-
-    // start the execution worker
-    let (mut manager, controller) = start_execution_worker(
-        exec_cfg.clone(),
-        sample_state.clone(),
-        sample_state.read().pos_state.selector.clone(),
-        mip_store,
-        channels,
-        Arc::new(RwLock::new(create_test_wallet(Some(PreHashMap::default())))),
-        MassaMetrics::new(
-            false,
-            "0.0.0.0:9898".parse().unwrap(),
-            32,
-            std::time::Duration::from_secs(5),
-        )
-        .0,
-    );
-    // initialize the execution system with genesis blocks
-    init_execution_worker(&exec_cfg, &storage, controller.clone());
-    // keypair associated to thread 0
+    let block_producer = KeyPair::generate(0).unwrap();
     let keypair = KeyPair::from_str(TEST_SK_1).unwrap();
+
+    let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
+    );
+    let mut universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg);
     // load bytecodes
     // you can check the source code of the following wasm files in massa-unit-tests-src
     let op_bytecode = include_bytes!("./wasm/deploy_sc.wasm");
@@ -874,17 +704,22 @@ fn sc_deployment() {
     datastore.insert(b"smart-contract".to_vec(), datastore_bytecode);
 
     // create the block containing the operation
-    let op = create_execute_sc_operation(&keypair, op_bytecode, datastore.clone()).unwrap();
-    storage.store_operations(vec![op.clone()]);
-    let block = create_block(
-        KeyPair::generate(0).unwrap(),
-        vec![op],
-        vec![],
-        Slot::new(1, 0),
+    let op = ExecutionTestUniverse::create_execute_sc_operation(
+        &keypair,
+        op_bytecode,
+        datastore.clone(),
     )
     .unwrap();
+    universe.storage.store_operations(vec![op.clone()]);
+    let block = ExecutionTestUniverse::create_block(
+        &block_producer,
+        Slot::new(1, 0),
+        vec![op],
+        vec![],
+        vec![],
+    );
     // store the block in storage
-    storage.store_block(block.clone());
+    universe.storage.store_block(block.clone());
 
     // set our block as a final block so the message is sent
     let mut finalized_blocks: HashMap<Slot, BlockId> = Default::default();
@@ -893,11 +728,11 @@ fn sc_deployment() {
     block_metadata.insert(
         block.id,
         ExecutionBlockMetadata {
-            same_thread_parent_creator: Some(get_random_address()),
-            storage: Some(storage),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
+            storage: Some(universe.storage.clone()),
         },
     );
-    controller.update_blockclique_status(
+    universe.module_controller.update_blockclique_status(
         finalized_blocks,
         Default::default(),
         block_metadata.clone(),
@@ -906,9 +741,11 @@ fn sc_deployment() {
     std::thread::sleep(Duration::from_millis(100));
 
     // retrieve events emitted by smart contracts
-    let events = controller.get_filtered_sc_output_event(EventFilter {
-        ..Default::default()
-    });
+    let events = universe
+        .module_controller
+        .get_filtered_sc_output_event(EventFilter {
+            ..Default::default()
+        });
 
     // match the events
     if events.len() != 3 {
@@ -920,9 +757,6 @@ fn sc_deployment() {
     assert_eq!(events[0].data, "sc created");
     assert_eq!(events[1].data, "constructor exists and will be called");
     assert_eq!(events[2].data, "constructor called by deployer");
-
-    // stop the execution controller
-    manager.stop();
 }
 
 /// # Context
@@ -944,53 +778,22 @@ fn sc_deployment() {
 #[test]
 #[serial]
 fn send_and_receive_async_message_with_trigger() {
-    // setup the period duration and the maximum gas for asynchronous messages execution
+    // setup the period duration
     let exec_cfg = ExecutionConfig {
         t0: MassaTime::from_millis(100),
         cursor_delay: MassaTime::from_millis(0),
         max_async_gas: 1_000_000_000,
         ..ExecutionConfig::default()
     };
-    // get a sample final state
-    let (sample_state, _keep_file, _keep_dir) = get_sample_state(0).unwrap();
-
-    // init the MIP store
-    let mip_stats_config = MipStatsConfig {
-        block_count_considered: MIP_STORE_STATS_BLOCK_CONSIDERED,
-        warn_announced_version_ratio: Ratio::new_raw(30, 100),
-    };
-    let mip_store = MipStore::try_from(([], mip_stats_config)).unwrap();
-
-    let mut blockclique_blocks: HashMap<Slot, BlockId> = HashMap::new();
-    // init the storage
-    let mut storage = Storage::create_root();
-
-    let slot_execution_output_sender = broadcast::channel(5000).0;
-
-    let channels = ExecutionChannels {
-        slot_execution_output_sender,
-    };
-
-    // start the execution worker
-    let (mut manager, controller) = start_execution_worker(
-        exec_cfg.clone(),
-        sample_state.clone(),
-        sample_state.read().pos_state.selector.clone(),
-        mip_store,
-        channels,
-        Arc::new(RwLock::new(create_test_wallet(Some(PreHashMap::default())))),
-        MassaMetrics::new(
-            false,
-            "0.0.0.0:9898".parse().unwrap(),
-            32,
-            std::time::Duration::from_secs(5),
-        )
-        .0,
-    );
-    // initialize the execution system with genesis blocks
-    init_execution_worker(&exec_cfg, &storage, controller.clone());
-    // keypair associated to thread 0
+    let block_producer = KeyPair::generate(0).unwrap();
     let keypair = KeyPair::from_str(TEST_SK_1).unwrap();
+    let mut blockclique_blocks: HashMap<Slot, BlockId> = HashMap::default();
+    let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
+    );
+    let mut universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg);
     // load bytecode
     // you can check the source code of the following wasm file in massa-unit-tests-src
     let bytecode = include_bytes!("./wasm/send_message_deploy_condition.wasm");
@@ -1007,11 +810,18 @@ fn send_and_receive_async_message_with_trigger() {
     datastore.insert(key, datastore_bytecode);
 
     // create the block containing the smart contract execution operation
-    let operation = create_execute_sc_operation(&keypair, bytecode, datastore).unwrap();
-    storage.store_operations(vec![operation.clone()]);
-    let block = create_block(keypair, vec![operation], vec![], Slot::new(1, 0)).unwrap();
+    let operation =
+        ExecutionTestUniverse::create_execute_sc_operation(&keypair, bytecode, datastore).unwrap();
+    universe.storage.store_operations(vec![operation.clone()]);
+    let block = ExecutionTestUniverse::create_block(
+        &block_producer,
+        Slot::new(1, 0),
+        vec![operation],
+        vec![],
+        vec![],
+    );
     // store the block in storage
-    storage.store_block(block.clone());
+    universe.storage.store_block(block.clone());
 
     // set our block as a final block so the message is sent
     let mut finalized_blocks: HashMap<Slot, BlockId> = Default::default();
@@ -1020,12 +830,12 @@ fn send_and_receive_async_message_with_trigger() {
     block_metadata.insert(
         block.id,
         ExecutionBlockMetadata {
-            same_thread_parent_creator: Some(get_random_address()),
-            storage: Some(storage.clone()),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
+            storage: Some(universe.storage.clone()),
         },
     );
     blockclique_blocks.insert(block.content.header.content.slot, block.id);
-    controller.update_blockclique_status(
+    universe.module_controller.update_blockclique_status(
         finalized_blocks.clone(),
         Some(blockclique_blocks.clone()),
         block_metadata.clone(),
@@ -1034,9 +844,11 @@ fn send_and_receive_async_message_with_trigger() {
     std::thread::sleep(Duration::from_millis(10));
 
     // retrieve events emitted by smart contracts
-    let events = controller.get_filtered_sc_output_event(EventFilter {
-        ..Default::default()
-    });
+    let events = universe
+        .module_controller
+        .get_filtered_sc_output_event(EventFilter {
+            ..Default::default()
+        });
 
     // match the events
     assert_eq!(events.len(), 2, "2 events were expected");
@@ -1050,11 +862,18 @@ fn send_and_receive_async_message_with_trigger() {
     let datastore = BTreeMap::new();
 
     // create the block containing the smart contract execution operation
-    let operation = create_execute_sc_operation(&keypair, bytecode, datastore).unwrap();
-    storage.store_operations(vec![operation.clone()]);
-    let block = create_block(keypair, vec![operation], vec![], Slot::new(1, 1)).unwrap();
+    let operation =
+        ExecutionTestUniverse::create_execute_sc_operation(&keypair, bytecode, datastore).unwrap();
+    universe.storage.store_operations(vec![operation.clone()]);
+    let block = ExecutionTestUniverse::create_block(
+        &block_producer,
+        Slot::new(1, 1),
+        vec![operation],
+        vec![],
+        vec![],
+    );
     // store the block in storage
-    storage.store_block(block.clone());
+    universe.storage.store_block(block.clone());
 
     // set our block as a final block so the message is sent
     finalized_blocks.insert(block.content.header.content.slot, block.id);
@@ -1062,19 +881,25 @@ fn send_and_receive_async_message_with_trigger() {
     block_metadata.insert(
         block.id,
         ExecutionBlockMetadata {
-            same_thread_parent_creator: Some(get_random_address()),
-            storage: Some(storage.clone()),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
+            storage: Some(universe.storage.clone()),
         },
     );
     blockclique_blocks.insert(block.content.header.content.slot, block.id);
-    controller.update_blockclique_status(finalized_blocks.clone(), None, block_metadata.clone());
+    universe.module_controller.update_blockclique_status(
+        finalized_blocks.clone(),
+        None,
+        block_metadata.clone(),
+    );
     // sleep for 10ms to reach the message execution period
     std::thread::sleep(Duration::from_millis(10));
 
     // retrieve events emitted by smart contracts
-    let events = controller.get_filtered_sc_output_event(EventFilter {
-        ..Default::default()
-    });
+    let events = universe
+        .module_controller
+        .get_filtered_sc_output_event(EventFilter {
+            ..Default::default()
+        });
 
     // match the events
     assert!(events.len() == 3, "3 events were expected");
@@ -1087,11 +912,18 @@ fn send_and_receive_async_message_with_trigger() {
     let bytecode = include_bytes!("./wasm/send_message_trigger.wasm");
     let datastore = BTreeMap::new();
 
-    let operation = create_execute_sc_operation(&keypair, bytecode, datastore).unwrap();
-    storage.store_operations(vec![operation.clone()]);
-    let block = create_block(keypair, vec![operation], vec![], Slot::new(1, 2)).unwrap();
+    let operation =
+        ExecutionTestUniverse::create_execute_sc_operation(&keypair, bytecode, datastore).unwrap();
+    universe.storage.store_operations(vec![operation.clone()]);
+    let block = ExecutionTestUniverse::create_block(
+        &block_producer,
+        Slot::new(1, 2),
+        vec![operation],
+        vec![],
+        vec![],
+    );
     // store the block in storage
-    storage.store_block(block.clone());
+    universe.storage.store_block(block.clone());
 
     // set our block as a final block so the message is sent
     finalized_blocks.insert(block.content.header.content.slot, block.id);
@@ -1099,76 +931,50 @@ fn send_and_receive_async_message_with_trigger() {
     block_metadata.insert(
         block.id,
         ExecutionBlockMetadata {
-            same_thread_parent_creator: Some(get_random_address()),
-            storage: Some(storage),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
+            storage: Some(universe.storage.clone()),
         },
     );
     blockclique_blocks.insert(block.content.header.content.slot, block.id);
-    controller.update_blockclique_status(finalized_blocks.clone(), None, block_metadata.clone());
+    universe.module_controller.update_blockclique_status(
+        finalized_blocks.clone(),
+        None,
+        block_metadata.clone(),
+    );
     // sleep for 1000ms to reach the message execution period
     std::thread::sleep(Duration::from_millis(1000));
 
     // retrieve events emitted by smart contracts
-    let events = controller.get_filtered_sc_output_event(EventFilter {
-        ..Default::default()
-    });
+    let events = universe
+        .module_controller
+        .get_filtered_sc_output_event(EventFilter {
+            ..Default::default()
+        });
 
     // match the events
     assert!(events.len() == 4, "4 events were expected");
-
-    manager.stop();
 }
 
 #[test]
 #[serial]
-pub fn send_and_receive_transaction() {
+fn send_and_receive_transaction() {
     // setup the period duration
     let exec_cfg = ExecutionConfig {
         t0: MassaTime::from_millis(100),
         cursor_delay: MassaTime::from_millis(0),
         ..ExecutionConfig::default()
     };
-    // get a sample final state
-    let (sample_state, _keep_file, _keep_dir) = get_sample_state(0).unwrap();
+    let block_producer = KeyPair::generate(0).unwrap();
+    let keypair = KeyPair::from_str(TEST_SK_1).unwrap();
 
-    // init the MIP store
-    let mip_stats_config = MipStatsConfig {
-        block_count_considered: MIP_STORE_STATS_BLOCK_CONSIDERED,
-        warn_announced_version_ratio: Ratio::new_raw(30, 100),
-    };
-    let mip_store = MipStore::try_from(([], mip_stats_config)).unwrap();
-
-    // init the storage
-    let mut storage = Storage::create_root();
-
-    let slot_execution_output_sender = broadcast::channel(5000).0;
-
-    let channels = ExecutionChannels {
-        slot_execution_output_sender,
-    };
-
-    // start the execution worker
-    let (mut manager, controller) = start_execution_worker(
-        exec_cfg.clone(),
-        sample_state.clone(),
-        sample_state.read().pos_state.selector.clone(),
-        mip_store,
-        channels,
-        Arc::new(RwLock::new(create_test_wallet(Some(PreHashMap::default())))),
-        MassaMetrics::new(
-            false,
-            "0.0.0.0:9898".parse().unwrap(),
-            32,
-            std::time::Duration::from_secs(5),
-        )
-        .0,
+    let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
     );
-    // initialize the execution system with genesis blocks
-    init_execution_worker(&exec_cfg, &storage, controller.clone());
-    // generate the sender_keypair and recipient_address
-
-    let sender_keypair = KeyPair::from_str(TEST_SK_1).unwrap();
-    let (recipient_address, _keypair) = get_random_address_full();
+    let mut universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg);
+    let recipient_address =
+        Address::from_public_key(&KeyPair::from_str(TEST_SK_2).unwrap().get_public_key());
 
     // create the operation
     let operation = Operation::new_verifiable(
@@ -1181,20 +987,20 @@ pub fn send_and_receive_transaction() {
             },
         },
         OperationSerializer::new(),
-        &sender_keypair,
+        &keypair,
     )
     .unwrap();
     // create the block containing the transaction operation
-    storage.store_operations(vec![operation.clone()]);
-    let block = create_block(
-        KeyPair::generate(0).unwrap(),
+    universe.storage.store_operations(vec![operation.clone()]);
+    let block = ExecutionTestUniverse::create_block(
+        &block_producer,
+        Slot::new(1, 0),
         vec![operation],
         vec![],
-        Slot::new(1, 0),
-    )
-    .unwrap();
+        vec![],
+    );
     // store the block in storage
-    storage.store_block(block.clone());
+    universe.storage.store_block(block.clone());
     // set our block as a final block so the transaction is processed
     let mut finalized_blocks: HashMap<Slot, BlockId> = Default::default();
     finalized_blocks.insert(block.content.header.content.slot, block.id);
@@ -1202,19 +1008,21 @@ pub fn send_and_receive_transaction() {
     block_metadata.insert(
         block.id,
         ExecutionBlockMetadata {
-            same_thread_parent_creator: Some(get_random_address()),
-            storage: Some(storage),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
+            storage: Some(universe.storage.clone()),
         },
     );
-    controller.update_blockclique_status(
+    universe.module_controller.update_blockclique_status(
         finalized_blocks,
         Default::default(),
         block_metadata.clone(),
     );
     std::thread::sleep(Duration::from_millis(10));
     // check recipient balance
+    //TODO: replace when ledger will be mocked
     assert_eq!(
-        sample_state
+        universe
+            .final_state
             .read()
             .ledger
             .get_balance(&recipient_address)
@@ -1225,59 +1033,27 @@ pub fn send_and_receive_transaction() {
             // Storage cost base
             .saturating_sub(LEDGER_ENTRY_BASE_COST)
     );
-    // stop the execution controller
-    manager.stop();
 }
 
 #[test]
 #[serial]
-pub fn roll_buy() {
+fn roll_buy() {
     // setup the period duration
     let exec_cfg = ExecutionConfig {
         t0: MassaTime::from_millis(100),
         cursor_delay: MassaTime::from_millis(0),
         ..ExecutionConfig::default()
     };
-    // get a sample final state
-    let (sample_state, _keep_file, _keep_dir) = get_sample_state(0).unwrap();
-
-    // init the MIP store
-    let mip_stats_config = MipStatsConfig {
-        block_count_considered: MIP_STORE_STATS_BLOCK_CONSIDERED,
-        warn_announced_version_ratio: Ratio::new_raw(30, 100),
-    };
-    let mip_store = MipStore::try_from(([], mip_stats_config)).unwrap();
-
-    // init the storage
-    let mut storage = Storage::create_root();
-
-    let slot_execution_output_sender = broadcast::channel(5000).0;
-
-    let channels = ExecutionChannels {
-        slot_execution_output_sender,
-    };
-
-    // start the execution worker
-    let (mut manager, controller) = start_execution_worker(
-        exec_cfg.clone(),
-        sample_state.clone(),
-        sample_state.read().pos_state.selector.clone(),
-        mip_store,
-        channels,
-        Arc::new(RwLock::new(create_test_wallet(Some(PreHashMap::default())))),
-        MassaMetrics::new(
-            false,
-            "0.0.0.0:9898".parse().unwrap(),
-            32,
-            std::time::Duration::from_secs(5),
-        )
-        .0,
-    );
-    // initialize the execution system with genesis blocks
-    init_execution_worker(&exec_cfg, &storage, controller.clone());
-    // generate the keypair and its corresponding address
+    let block_producer = KeyPair::generate(0).unwrap();
     let keypair = KeyPair::from_str(TEST_SK_1).unwrap();
     let address = Address::from_public_key(&keypair.get_public_key());
+
+    let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
+    );
+    let mut universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg);
     // create the operation
     let operation = Operation::new_verifiable(
         Operation {
@@ -1290,16 +1066,16 @@ pub fn roll_buy() {
     )
     .unwrap();
     // create the block containing the roll buy operation
-    storage.store_operations(vec![operation.clone()]);
-    let block = create_block(
-        KeyPair::generate(0).unwrap(),
+    universe.storage.store_operations(vec![operation.clone()]);
+    let block = ExecutionTestUniverse::create_block(
+        &block_producer,
+        Slot::new(1, 0),
         vec![operation],
         vec![],
-        Slot::new(1, 0),
-    )
-    .unwrap();
+        vec![],
+    );
     // store the block in storage
-    storage.store_block(block.clone());
+    universe.storage.store_block(block.clone());
     // set our block as a final block so the purchase is processed
     let mut finalized_blocks: HashMap<Slot, BlockId> = Default::default();
     finalized_blocks.insert(block.content.header.content.slot, block.id);
@@ -1307,93 +1083,70 @@ pub fn roll_buy() {
     block_metadata.insert(
         block.id,
         ExecutionBlockMetadata {
-            same_thread_parent_creator: Some(get_random_address()),
-            storage: Some(storage),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
+            storage: Some(universe.storage.clone()),
         },
     );
-    controller.update_blockclique_status(
+    universe.module_controller.update_blockclique_status(
         finalized_blocks,
         Default::default(),
         block_metadata.clone(),
     );
     std::thread::sleep(Duration::from_millis(100));
     // check roll count of the buyer address and its balance
-    let sample_read = sample_state.read();
+    let sample_read = universe.final_state.read();
     assert_eq!(sample_read.pos_state.get_rolls_for(&address), 110);
     assert_eq!(
         sample_read.ledger.get_balance(&address).unwrap(),
         Amount::from_str("299_000").unwrap()
     );
-    // stop the execution controller
-    manager.stop();
 }
 
 #[test]
 #[serial]
-pub fn roll_sell() {
+fn roll_sell() {
     // Try to sell 10 rolls (operation 1) then 1 rolls (operation 2)
     // Check for resulting roll count + resulting deferred credits
+    // setup the period duration
 
     // setup the period duration
-    let mut exec_cfg = ExecutionConfig {
+    // turn off roll selling on missed block opportunities
+    // otherwise balance will be credited with those sold roll (and we need to check the balance for
+    // if the deferred credits are reimbursed
+    let exec_cfg = ExecutionConfig {
         t0: MassaTime::from_millis(100),
         cursor_delay: MassaTime::from_millis(0),
         periods_per_cycle: 2,
         thread_count: 2,
         last_start_period: 2,
+        max_miss_ratio: Ratio::new(1, 1),
         ..Default::default()
     };
-    // turn off roll selling on missed block opportunities
-    // otherwise balance will be credited with those sold roll (and we need to check the balance for
-    // if the deferred credits are reimbursed
-    exec_cfg.max_miss_ratio = Ratio::new(1, 1);
-
-    // get a sample final state
-    let (sample_state, _keep_file, _keep_dir) = get_sample_state(2).unwrap();
-
-    // init the MIP store
-    let mip_stats_config = MipStatsConfig {
-        block_count_considered: MIP_STORE_STATS_BLOCK_CONSIDERED,
-        warn_announced_version_ratio: Ratio::new_raw(30, 100),
-    };
-    let mip_store = MipStore::try_from(([], mip_stats_config)).unwrap();
-
-    // init the storage
-    let mut storage = Storage::create_root();
-
-    let slot_execution_output_sender = broadcast::channel(5000).0;
-
-    let channels = ExecutionChannels {
-        slot_execution_output_sender,
-    };
-
-    // start the execution worker
-    let (mut manager, controller) = start_execution_worker(
-        exec_cfg.clone(),
-        sample_state.clone(),
-        sample_state.read().pos_state.selector.clone(),
-        mip_store,
-        channels,
-        Arc::new(RwLock::new(create_test_wallet(Some(PreHashMap::default())))),
-        MassaMetrics::new(
-            false,
-            "0.0.0.0:9898".parse().unwrap(),
-            32,
-            std::time::Duration::from_secs(5),
-        )
-        .0,
-    );
-    // initialize the execution system with genesis blocks
-    init_execution_worker(&exec_cfg, &storage, controller.clone());
-    // generate the keypair and its corresponding address
+    let block_producer = KeyPair::generate(0).unwrap();
     let keypair = KeyPair::from_str(TEST_SK_1).unwrap();
     let address = Address::from_public_key(&keypair.get_public_key());
 
+    let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
+    );
+    let mut universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg.clone());
+
     // get initial balance
-    let balance_initial = sample_state.read().ledger.get_balance(&address).unwrap();
+    let balance_initial = universe
+        .final_state
+        .read()
+        .ledger
+        .get_balance(&address)
+        .unwrap();
 
     // get initial roll count
-    let roll_count_initial = sample_state.read().pos_state.get_rolls_for(&address);
+    let roll_count_initial = universe
+        .final_state
+        .read()
+        .pos_state
+        .get_rolls_for(&address);
     let roll_sell_1 = 10;
     let roll_sell_2 = 1;
 
@@ -1402,14 +1155,19 @@ pub fn roll_sell() {
     let mut batch = DBBatch::new();
 
     // set initial_deferred_credits that will be reimbursed at first block
-    sample_state.write().pos_state.put_deferred_credits_entry(
-        &Slot::new(1, 0),
-        &address,
-        &initial_deferred_credits,
-        &mut batch,
-    );
+    universe
+        .final_state
+        .write()
+        .pos_state
+        .put_deferred_credits_entry(
+            &Slot::new(1, 0),
+            &address,
+            &initial_deferred_credits,
+            &mut batch,
+        );
 
-    sample_state
+    universe
+        .final_state
         .write()
         .db
         .write()
@@ -1441,16 +1199,18 @@ pub fn roll_sell() {
     )
     .unwrap();
     // create the block containing the roll buy operation
-    storage.store_operations(vec![operation1.clone(), operation2.clone()]);
-    let block = create_block(
-        KeyPair::generate(0).unwrap(),
+    universe
+        .storage
+        .store_operations(vec![operation1.clone(), operation2.clone()]);
+    let block = ExecutionTestUniverse::create_block(
+        &block_producer,
+        Slot::new(3, 0),
         vec![operation1, operation2],
         vec![],
-        Slot::new(3, 0),
-    )
-    .unwrap();
+        vec![],
+    );
     // store the block in storage
-    storage.store_block(block.clone());
+    universe.storage.store_block(block.clone());
     // set the block as final so the sell and credits are processed
     let mut finalized_blocks: HashMap<Slot, BlockId> = Default::default();
     finalized_blocks.insert(block.content.header.content.slot, block.id);
@@ -1458,11 +1218,11 @@ pub fn roll_sell() {
     block_metadata.insert(
         block.id,
         ExecutionBlockMetadata {
-            same_thread_parent_creator: Some(get_random_address()),
-            storage: Some(storage),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
+            storage: Some(universe.storage.clone()),
         },
     );
-    controller.update_blockclique_status(
+    universe.module_controller.update_blockclique_status(
         finalized_blocks,
         Default::default(),
         block_metadata.clone(),
@@ -1470,7 +1230,7 @@ pub fn roll_sell() {
     std::thread::sleep(Duration::from_millis(1000));
 
     // check roll count deferred credits and candidate balance of the seller address
-    let sample_read = sample_state.read();
+    let sample_read = universe.final_state.read();
     let mut credits = PreHashMap::default();
     let roll_remaining = roll_count_initial - roll_sell_1 - roll_sell_2;
     let roll_sold = roll_sell_1 + roll_sell_2;
@@ -1522,7 +1282,9 @@ pub fn roll_sell() {
     );
 
     // Now check balance
-    let balances = controller.get_final_and_candidate_balance(&[address]);
+    let balances = universe
+        .module_controller
+        .get_final_and_candidate_balance(&[address]);
     let candidate_balance = balances.get(0).unwrap().1.unwrap();
 
     assert_eq!(
@@ -1536,81 +1298,54 @@ pub fn roll_sell() {
             .checked_add(initial_deferred_credits)
             .unwrap()
     );
-
-    // stop the execution controller
-    manager.stop();
 }
 
 #[test]
 #[serial]
-pub fn roll_slash() {
+fn roll_slash() {
     // Try to sell 97 rolls (operation 1) then process a Denunciation (with config set to slash
     // 3 rolls)
     // Check for resulting roll & deferred credits & balance
 
     // setup the period duration
-    let mut exec_cfg = ExecutionConfig {
+    // turn off roll selling on missed block opportunities
+    // otherwise balance will be credited with those sold roll (and we need to check the balance for
+    // if the deferred credits are reimbursed
+    let exec_cfg = ExecutionConfig {
         t0: MassaTime::from_millis(100),
         cursor_delay: MassaTime::from_millis(0),
         periods_per_cycle: 2,
         thread_count: 2,
         last_start_period: 2,
         roll_count_to_slash_on_denunciation: 3, // Set to 3 to check if config is taken into account
+        max_miss_ratio: Ratio::new(1, 1),
         ..Default::default()
     };
-    // turn off roll selling on missed block opportunities
-    // otherwise balance will be credited with those sold roll (and we need to check the balance for
-    // if the deferred credits are reimbursed
-    exec_cfg.max_miss_ratio = Ratio::new(1, 1);
-
-    // get a sample final state
-    let (sample_state, _keep_file, _keep_dir) =
-        get_sample_state(exec_cfg.last_start_period).unwrap();
-
-    // init the MIP store
-    let mip_stats_config = MipStatsConfig {
-        block_count_considered: MIP_STORE_STATS_BLOCK_CONSIDERED,
-        warn_announced_version_ratio: Ratio::new_raw(30, 100),
-    };
-    let mip_store = MipStore::try_from(([], mip_stats_config)).unwrap();
-
-    // init the storage
-    let mut storage = Storage::create_root();
-
-    let slot_execution_output_sender = broadcast::channel(5000).0;
-
-    let channels = ExecutionChannels {
-        slot_execution_output_sender,
-    };
-
-    // start the execution worker
-    let (mut manager, controller) = start_execution_worker(
-        exec_cfg.clone(),
-        sample_state.clone(),
-        sample_state.read().pos_state.selector.clone(),
-        mip_store,
-        channels,
-        Arc::new(RwLock::new(create_test_wallet(Some(PreHashMap::default())))),
-        MassaMetrics::new(
-            false,
-            "0.0.0.0:9898".parse().unwrap(),
-            32,
-            std::time::Duration::from_secs(5),
-        )
-        .0,
-    );
-    // initialize the execution system with genesis blocks
-    init_execution_worker(&exec_cfg, &storage, controller.clone());
-
-    // generate the keypair and its corresponding address
+    let block_producer = KeyPair::generate(0).unwrap();
     let keypair = KeyPair::from_str(TEST_SK_1).unwrap();
     let address = Address::from_public_key(&keypair.get_public_key());
 
+    let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
+    );
+    let mut universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg.clone());
+
     // get initial balance
-    let balance_initial = sample_state.read().ledger.get_balance(&address).unwrap();
+    let balance_initial = universe
+        .final_state
+        .read()
+        .ledger
+        .get_balance(&address)
+        .unwrap();
 
     // get initial roll count
-    let roll_count_initial = sample_state.read().pos_state.get_rolls_for(&address);
+    let roll_count_initial = universe
+        .final_state
+        .read()
+        .pos_state
+        .get_rolls_for(&address);
     let roll_to_sell = roll_count_initial
         .checked_sub(exec_cfg.roll_count_to_slash_on_denunciation)
         .unwrap();
@@ -1637,21 +1372,21 @@ pub fn roll_slash() {
     // create a denunciation (that will be ignored as it has been created at the last start period)
     let (_slot, _keypair, s_endorsement_1, s_endorsement_2, _) = gen_endorsements_for_denunciation(
         Some(Slot::new(exec_cfg.last_start_period, 4)),
-        Some(keypair),
+        Some(keypair.clone()),
     );
     let denunciation_2 = Denunciation::try_from((&s_endorsement_1, &s_endorsement_2)).unwrap();
 
     // create the block containing the roll buy operation
-    storage.store_operations(vec![operation1.clone()]);
-    let block = create_block(
-        KeyPair::generate(0).unwrap(),
-        vec![operation1],
-        vec![denunciation.clone(), denunciation, denunciation_2],
+    universe.storage.store_operations(vec![operation1.clone()]);
+    let block = ExecutionTestUniverse::create_block(
+        &block_producer,
         Slot::new(3, 0),
-    )
-    .unwrap();
+        vec![operation1],
+        vec![],
+        vec![denunciation.clone(), denunciation, denunciation_2],
+    );
     // store the block in storage
-    storage.store_block(block.clone());
+    universe.storage.store_block(block.clone());
     // set the block as final so the sell and credits are processed
     let mut finalized_blocks: HashMap<Slot, BlockId> = Default::default();
     finalized_blocks.insert(block.content.header.content.slot, block.id);
@@ -1659,11 +1394,11 @@ pub fn roll_slash() {
     block_metadata.insert(
         block.id,
         ExecutionBlockMetadata {
-            same_thread_parent_creator: Some(get_random_address()),
-            storage: Some(storage),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
+            storage: Some(universe.storage.clone()),
         },
     );
-    controller.update_blockclique_status(
+    universe.module_controller.update_blockclique_status(
         finalized_blocks,
         Default::default(),
         block_metadata.clone(),
@@ -1671,7 +1406,7 @@ pub fn roll_slash() {
     std::thread::sleep(Duration::from_millis(1000));
 
     // check roll count deferred credits and candidate balance of the seller address
-    let sample_read = sample_state.read();
+    let sample_read = universe.final_state.read();
     let mut credits = PreHashMap::default();
     let roll_sold = roll_to_sell;
     credits.insert(
@@ -1696,7 +1431,9 @@ pub fn roll_slash() {
     );
 
     // Now check balance
-    let balances = controller.get_final_and_candidate_balance(&[address]);
+    let balances = universe
+        .module_controller
+        .get_final_and_candidate_balance(&[address]);
     let candidate_balance = balances.get(0).unwrap().1.unwrap();
 
     assert_eq!(
@@ -1708,81 +1445,54 @@ pub fn roll_slash() {
             .checked_add(balance_initial)
             .unwrap()
     );
-
-    // stop the execution controller
-    manager.stop();
 }
 
 #[test]
 #[serial]
-pub fn roll_slash_2() {
+fn roll_slash_2() {
     // Try to sell all rolls (operation 1) then process a Denunciation (with config set to slash
     // 4 rolls)
     // Check for resulting roll & deferred credits & balance
 
     // setup the period duration
-    let mut exec_cfg = ExecutionConfig {
+    // turn off roll selling on missed block opportunities
+    // otherwise balance will be credited with those sold roll (and we need to check the balance for
+    // if the deferred credits are reimbursed
+    let exec_cfg = ExecutionConfig {
         t0: MassaTime::from_millis(100),
         cursor_delay: MassaTime::from_millis(0),
         periods_per_cycle: 2,
         thread_count: 2,
         last_start_period: 2,
-        roll_count_to_slash_on_denunciation: 4, // Set to 4 to check if config is taken into account
+        roll_count_to_slash_on_denunciation: 3, // Set to 3 to check if config is taken into account
+        max_miss_ratio: Ratio::new(1, 1),
         ..Default::default()
     };
-    // turn off roll selling on missed block opportunities
-    // otherwise balance will be credited with those sold roll (and we need to check the balance for
-    // if the deferred credits are reimbursed
-    exec_cfg.max_miss_ratio = Ratio::new(1, 1);
-
-    // get a sample final state
-    let (sample_state, _keep_file, _keep_dir) =
-        get_sample_state(exec_cfg.last_start_period).unwrap();
-
-    // init the MIP store
-    let mip_stats_config = MipStatsConfig {
-        block_count_considered: MIP_STORE_STATS_BLOCK_CONSIDERED,
-        warn_announced_version_ratio: Ratio::new_raw(30, 100),
-    };
-    let mip_store = MipStore::try_from(([], mip_stats_config)).unwrap();
-
-    // init the storage
-    let mut storage = Storage::create_root();
-
-    let slot_execution_output_sender = broadcast::channel(5000).0;
-
-    let channels = ExecutionChannels {
-        slot_execution_output_sender,
-    };
-
-    // start the execution worker
-    let (mut manager, controller) = start_execution_worker(
-        exec_cfg.clone(),
-        sample_state.clone(),
-        sample_state.read().pos_state.selector.clone(),
-        mip_store,
-        channels,
-        Arc::new(RwLock::new(create_test_wallet(Some(PreHashMap::default())))),
-        MassaMetrics::new(
-            false,
-            "0.0.0.0:9898".parse().unwrap(),
-            32,
-            std::time::Duration::from_secs(5),
-        )
-        .0,
-    );
-    // initialize the execution system with genesis blocks
-    init_execution_worker(&exec_cfg, &storage, controller.clone());
-
-    // generate the keypair and its corresponding address
+    let block_producer = KeyPair::generate(0).unwrap();
     let keypair = KeyPair::from_str(TEST_SK_1).unwrap();
     let address = Address::from_public_key(&keypair.get_public_key());
 
+    let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
+    );
+    let mut universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg.clone());
+
     // get initial balance
-    let balance_initial = sample_state.read().ledger.get_balance(&address).unwrap();
+    let balance_initial = universe
+        .final_state
+        .read()
+        .ledger
+        .get_balance(&address)
+        .unwrap();
 
     // get initial roll count
-    let roll_count_initial = sample_state.read().pos_state.get_rolls_for(&address);
+    let roll_count_initial = universe
+        .final_state
+        .read()
+        .pos_state
+        .get_rolls_for(&address);
     // sell all rolls so we can check if slash will occur on deferred credits
     let roll_to_sell_1 = 1;
     let roll_to_sell_2 = roll_count_initial - 1;
@@ -1824,20 +1534,22 @@ pub fn roll_slash_2() {
 
     // create a denunciation
     let (_slot, _keypair, s_endorsement_1, s_endorsement_2, _) =
-        gen_endorsements_for_denunciation(Some(Slot::new(3, 0)), Some(keypair));
+        gen_endorsements_for_denunciation(Some(Slot::new(3, 0)), Some(keypair.clone()));
     let denunciation = Denunciation::try_from((&s_endorsement_1, &s_endorsement_2)).unwrap();
 
     // create the block containing the roll buy operation
-    storage.store_operations(vec![operation1.clone(), operation2.clone()]);
-    let block = create_block(
-        KeyPair::generate(0).unwrap(),
-        vec![operation1, operation2],
-        vec![denunciation.clone(), denunciation],
+    universe
+        .storage
+        .store_operations(vec![operation1.clone(), operation2.clone()]);
+    let block = ExecutionTestUniverse::create_block(
+        &block_producer,
         Slot::new(3, 0),
-    )
-    .unwrap();
+        vec![operation1, operation2],
+        vec![],
+        vec![denunciation.clone(), denunciation],
+    );
     // store the block in storage
-    storage.store_block(block.clone());
+    universe.storage.store_block(block.clone());
     // set the block as final so the sell and credits are processed
     let mut finalized_blocks: HashMap<Slot, BlockId> = Default::default();
     finalized_blocks.insert(block.content.header.content.slot, block.id);
@@ -1845,11 +1557,11 @@ pub fn roll_slash_2() {
     block_metadata.insert(
         block.id,
         ExecutionBlockMetadata {
-            same_thread_parent_creator: Some(get_random_address()),
-            storage: Some(storage),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
+            storage: Some(universe.storage.clone()),
         },
     );
-    controller.update_blockclique_status(
+    universe.module_controller.update_blockclique_status(
         finalized_blocks,
         Default::default(),
         block_metadata.clone(),
@@ -1857,7 +1569,7 @@ pub fn roll_slash_2() {
     std::thread::sleep(Duration::from_millis(1000));
 
     // check roll count & deferred credits & candidate balance
-    let sample_read = sample_state.read();
+    let sample_read = universe.final_state.read();
     let mut credits = PreHashMap::default();
     let roll_sold = roll_to_sell;
     credits.insert(
@@ -1887,7 +1599,9 @@ pub fn roll_slash_2() {
     );
 
     // Now check balance
-    let balances = controller.get_final_and_candidate_balance(&[address]);
+    let balances = universe
+        .module_controller
+        .get_final_and_candidate_balance(&[address]);
     let candidate_balance = balances.get(0).unwrap().1.unwrap();
 
     assert_eq!(
@@ -1901,9 +1615,6 @@ pub fn roll_slash_2() {
             .checked_add(balance_initial)
             .unwrap()
     );
-
-    // stop the execution controller
-    manager.stop();
 }
 
 #[test]
@@ -1916,60 +1627,32 @@ fn sc_execution_error() {
         max_async_gas: 100_000,
         ..ExecutionConfig::default()
     };
-    // get a sample final state
-    let (sample_state, _keep_file, _keep_dir) = get_sample_state(0).unwrap();
-
-    // init the MIP store
-    let mip_stats_config = MipStatsConfig {
-        block_count_considered: MIP_STORE_STATS_BLOCK_CONSIDERED,
-        warn_announced_version_ratio: Ratio::new_raw(30, 100),
-    };
-    let mip_store = MipStore::try_from(([], mip_stats_config)).unwrap();
-
-    // init the storage
-    let mut storage = Storage::create_root();
-
-    let slot_execution_output_sender = broadcast::channel(5000).0;
-
-    let channels = ExecutionChannels {
-        slot_execution_output_sender,
-    };
-
-    // start the execution worker
-    let (mut manager, controller) = start_execution_worker(
-        exec_cfg.clone(),
-        sample_state.clone(),
-        sample_state.read().pos_state.selector.clone(),
-        mip_store,
-        channels,
-        Arc::new(RwLock::new(create_test_wallet(Some(PreHashMap::default())))),
-        MassaMetrics::new(
-            false,
-            "0.0.0.0:9898".parse().unwrap(),
-            32,
-            std::time::Duration::from_secs(5),
-        )
-        .0,
-    );
-    // initialize the execution system with genesis blocks
-    init_execution_worker(&exec_cfg, &storage, controller.clone());
-    // keypair associated to thread 0
+    let block_producer = KeyPair::generate(0).unwrap();
     let keypair = KeyPair::from_str(TEST_SK_1).unwrap();
+
+    let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
+    );
+    let mut universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg.clone());
     // load bytecode
     // you can check the source code of the following wasm file in massa-unit-tests-src
     let bytecode = include_bytes!("./wasm/execution_error.wasm");
     // create the block containing the erroneous smart contract execution operation
-    let operation = create_execute_sc_operation(&keypair, bytecode, BTreeMap::default()).unwrap();
-    storage.store_operations(vec![operation.clone()]);
-    let block = create_block(
-        KeyPair::generate(0).unwrap(),
+    let operation =
+        ExecutionTestUniverse::create_execute_sc_operation(&keypair, bytecode, BTreeMap::default())
+            .unwrap();
+    universe.storage.store_operations(vec![operation.clone()]);
+    let block = ExecutionTestUniverse::create_block(
+        &keypair,
+        Slot::new(1, 0),
         vec![operation],
         vec![],
-        Slot::new(1, 0),
-    )
-    .unwrap();
+        vec![],
+    );
     // store the block in storage
-    storage.store_block(block.clone());
+    universe.storage.store_block(block.clone());
     // set our block as a final block
     let mut finalized_blocks: HashMap<Slot, BlockId> = Default::default();
     finalized_blocks.insert(block.content.header.content.slot, block.id);
@@ -1977,11 +1660,11 @@ fn sc_execution_error() {
     block_metadata.insert(
         block.id,
         ExecutionBlockMetadata {
-            same_thread_parent_creator: Some(get_random_address()),
-            storage: Some(storage),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
+            storage: Some(universe.storage.clone()),
         },
     );
-    controller.update_blockclique_status(
+    universe.module_controller.update_blockclique_status(
         finalized_blocks,
         Default::default(),
         block_metadata.clone(),
@@ -1989,10 +1672,12 @@ fn sc_execution_error() {
     std::thread::sleep(Duration::from_millis(100));
 
     // retrieve the event emitted by the execution error
-    let events = controller.get_filtered_sc_output_event(EventFilter {
-        is_error: Some(true),
-        ..Default::default()
-    });
+    let events = universe
+        .module_controller
+        .get_filtered_sc_output_event(EventFilter {
+            is_error: Some(true),
+            ..Default::default()
+        });
     // match the events
     assert!(!events.is_empty(), "2 events were expected");
     assert_eq!(events[0].data, "event generated before the sc failure");
@@ -2001,8 +1686,6 @@ fn sc_execution_error() {
         .data
         .contains("runtime error when executing operation"));
     assert!(events[1].data.contains("address parsing error"));
-    // stop the execution controller
-    manager.stop();
 }
 
 #[test]
@@ -2015,62 +1698,33 @@ fn sc_datastore() {
         max_async_gas: 100_000,
         ..ExecutionConfig::default()
     };
-    // get a sample final state
-    let (sample_state, _keep_file, _keep_dir) = get_sample_state(0).unwrap();
-
-    // init the MIP store
-    let mip_stats_config = MipStatsConfig {
-        block_count_considered: MIP_STORE_STATS_BLOCK_CONSIDERED,
-        warn_announced_version_ratio: Ratio::new_raw(30, 100),
-    };
-    let mip_store = MipStore::try_from(([], mip_stats_config)).unwrap();
-
-    // init the storage
-    let mut storage = Storage::create_root();
-
-    let slot_execution_output_sender = broadcast::channel(5000).0;
-
-    let channels = ExecutionChannels {
-        slot_execution_output_sender,
-    };
-
-    // start the execution worker
-    let (mut manager, controller) = start_execution_worker(
-        exec_cfg.clone(),
-        sample_state.clone(),
-        sample_state.read().pos_state.selector.clone(),
-        mip_store,
-        channels,
-        Arc::new(RwLock::new(create_test_wallet(Some(PreHashMap::default())))),
-        MassaMetrics::new(
-            false,
-            "0.0.0.0:9898".parse().unwrap(),
-            32,
-            std::time::Duration::from_secs(5),
-        )
-        .0,
-    );
-    // initialize the execution system with genesis blocks
-    init_execution_worker(&exec_cfg, &storage, controller.clone());
-    // keypair associated to thread 0
+    let block_producer = KeyPair::generate(0).unwrap();
     let keypair = KeyPair::from_str(TEST_SK_1).unwrap();
+
+    let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
+    );
+    let mut universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg.clone());
     // load bytecode
     // you can check the source code of the following wasm file in massa-unit-tests-src
     let bytecode = include_bytes!("./wasm/datastore.wasm");
     let datastore = BTreeMap::from([(vec![65, 66], vec![255]), (vec![9], vec![10, 11])]);
 
     // create the block containing the erroneous smart contract execution operation
-    let operation = create_execute_sc_operation(&keypair, bytecode, datastore).unwrap();
-    storage.store_operations(vec![operation.clone()]);
-    let block = create_block(
-        KeyPair::generate(0).unwrap(),
+    let operation =
+        ExecutionTestUniverse::create_execute_sc_operation(&keypair, bytecode, datastore).unwrap();
+    universe.storage.store_operations(vec![operation.clone()]);
+    let block = ExecutionTestUniverse::create_block(
+        &keypair,
+        Slot::new(1, 0),
         vec![operation],
         vec![],
-        Slot::new(1, 0),
-    )
-    .unwrap();
+        vec![],
+    );
     // store the block in storage
-    storage.store_block(block.clone());
+    universe.storage.store_block(block.clone());
     // set our block as a final block
     let mut finalized_blocks: HashMap<Slot, BlockId> = Default::default();
     finalized_blocks.insert(block.content.header.content.slot, block.id);
@@ -2078,11 +1732,11 @@ fn sc_datastore() {
     block_metadata.insert(
         block.id,
         ExecutionBlockMetadata {
-            same_thread_parent_creator: Some(get_random_address()),
-            storage: Some(storage),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
+            storage: Some(universe.storage.clone()),
         },
     );
-    controller.update_blockclique_status(
+    universe.module_controller.update_blockclique_status(
         finalized_blocks,
         Some(Default::default()),
         block_metadata,
@@ -2090,67 +1744,36 @@ fn sc_datastore() {
     std::thread::sleep(Duration::from_millis(100));
 
     // retrieve the event emitted by the execution error
-    let events = controller.get_filtered_sc_output_event(EventFilter::default());
+    let events = universe
+        .module_controller
+        .get_filtered_sc_output_event(EventFilter::default());
 
     // match the events
     assert_eq!(events.len(), 3);
     assert_eq!(events[0].data, "keys: 9,65,66");
     assert_eq!(events[1].data, "has_key_1: true - has_key_2: false");
     assert_eq!(events[2].data, "data key 1: 255 - data key 3: 10,11");
-
-    // stop the execution controller
-    manager.stop();
 }
 
 #[test]
 #[serial]
 fn set_bytecode_error() {
-    // setup the period duration and the maximum gas for asynchronous messages execution
+    // setup the period duration
     let exec_cfg = ExecutionConfig {
         t0: MassaTime::from_millis(100),
         cursor_delay: MassaTime::from_millis(0),
         max_async_gas: 100_000,
         ..ExecutionConfig::default()
     };
-    // get a sample final state
-    let (sample_state, _keep_file, _keep_dir) = get_sample_state(0).unwrap();
-
-    // init the MIP store
-    let mip_stats_config = MipStatsConfig {
-        block_count_considered: MIP_STORE_STATS_BLOCK_CONSIDERED,
-        warn_announced_version_ratio: Ratio::new_raw(30, 100),
-    };
-    let mip_store = MipStore::try_from(([], mip_stats_config)).unwrap();
-
-    // init the storage
-    let mut storage = Storage::create_root();
-
-    let slot_execution_output_sender = broadcast::channel(5000).0;
-
-    let channels = ExecutionChannels {
-        slot_execution_output_sender,
-    };
-
-    // start the execution worker
-    let (_manager, controller) = start_execution_worker(
-        exec_cfg.clone(),
-        sample_state.clone(),
-        sample_state.read().pos_state.selector.clone(),
-        mip_store,
-        channels,
-        Arc::new(RwLock::new(create_test_wallet(Some(PreHashMap::default())))),
-        MassaMetrics::new(
-            false,
-            "0.0.0.0:9898".parse().unwrap(),
-            32,
-            std::time::Duration::from_secs(5),
-        )
-        .0,
-    );
-    // initialize the execution system with genesis blocks
-    init_execution_worker(&exec_cfg, &storage, controller.clone());
-    // keypair associated to thread 0
+    let block_producer = KeyPair::generate(0).unwrap();
     let keypair = KeyPair::from_str(TEST_SK_1).unwrap();
+
+    let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
+    );
+    let mut universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg.clone());
     // load bytecodes
     // you can check the source code of the following wasm file in massa-unit-tests-src
     let bytecode = include_bytes!("./wasm/set_bytecode_fail.wasm");
@@ -2159,17 +1782,18 @@ fn set_bytecode_error() {
     datastore.insert(b"smart-contract".to_vec(), datastore_bytecode);
 
     // create the block containing the erroneous smart contract execution operation
-    let operation = create_execute_sc_operation(&keypair, bytecode, datastore).unwrap();
-    storage.store_operations(vec![operation.clone()]);
-    let block = create_block(
-        KeyPair::generate(0).unwrap(),
+    let operation =
+        ExecutionTestUniverse::create_execute_sc_operation(&keypair, bytecode, datastore).unwrap();
+    universe.storage.store_operations(vec![operation.clone()]);
+    let block = ExecutionTestUniverse::create_block(
+        &keypair,
+        Slot::new(1, 0),
         vec![operation],
         vec![],
-        Slot::new(1, 0),
-    )
-    .unwrap();
+        vec![],
+    );
     // store the block in storage
-    storage.store_block(block.clone());
+    universe.storage.store_block(block.clone());
     // set our block as a final block
     let mut finalized_blocks: HashMap<Slot, BlockId> = Default::default();
     finalized_blocks.insert(block.content.header.content.slot, block.id);
@@ -2177,11 +1801,11 @@ fn set_bytecode_error() {
     block_metadata.insert(
         block.id,
         ExecutionBlockMetadata {
-            same_thread_parent_creator: Some(get_random_address()),
-            storage: Some(storage),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
+            storage: Some(universe.storage.clone()),
         },
     );
-    controller.update_blockclique_status(
+    universe.module_controller.update_blockclique_status(
         finalized_blocks,
         Default::default(),
         block_metadata.clone(),
@@ -2189,7 +1813,9 @@ fn set_bytecode_error() {
     std::thread::sleep(Duration::from_millis(10));
 
     // retrieve the event emitted by the execution error
-    let events = controller.get_filtered_sc_output_event(EventFilter::default());
+    let events = universe
+        .module_controller
+        .get_filtered_sc_output_event(EventFilter::default());
     // match the events
     assert!(!events.is_empty(), "One event was expected");
     assert!(events[0].data.contains("massa_execution_error"));
@@ -2202,83 +1828,55 @@ fn set_bytecode_error() {
 #[test]
 #[serial]
 fn datastore_manipulations() {
-    // setup the period duration and the maximum gas for asynchronous messages execution
+    // setup the period duration
     let exec_cfg = ExecutionConfig {
         t0: MassaTime::from_millis(100),
         cursor_delay: MassaTime::from_millis(0),
-        max_async_gas: 100_000,
         ..ExecutionConfig::default()
     };
-    // get a sample final state
-    let (sample_state, _keep_file, _keep_dir) = get_sample_state(0).unwrap();
-
-    // init the MIP store
-    let mip_stats_config = MipStatsConfig {
-        block_count_considered: MIP_STORE_STATS_BLOCK_CONSIDERED,
-        warn_announced_version_ratio: Ratio::new_raw(30, 100),
-    };
-    let mip_store = MipStore::try_from(([], mip_stats_config)).unwrap();
-
-    // init the storage
-    let mut storage = Storage::create_root();
-
-    let slot_execution_output_sender = broadcast::channel(5000).0;
-
-    let channels = ExecutionChannels {
-        slot_execution_output_sender,
-    };
-
-    // start the execution worker
-    let (mut manager, controller) = start_execution_worker(
-        exec_cfg.clone(),
-        sample_state.clone(),
-        sample_state.read().pos_state.selector.clone(),
-        mip_store,
-        channels,
-        Arc::new(RwLock::new(create_test_wallet(Some(PreHashMap::default())))),
-        MassaMetrics::new(
-            false,
-            "0.0.0.0:9898".parse().unwrap(),
-            32,
-            std::time::Duration::from_secs(5),
-        )
-        .0,
-    );
-    // initialize the execution system with genesis blocks
-    init_execution_worker(&exec_cfg, &storage, controller.clone());
-
-    // keypair associated to thread 0
+    let block_producer = KeyPair::generate(0).unwrap();
     let keypair = KeyPair::from_str(TEST_SK_1).unwrap();
-    // let address = Address::from_public_key(&keypair.get_public_key());
 
+    let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
+    );
+    let mut universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg.clone());
     // load bytecode
     // you can check the source code of the following wasm file in massa-unit-tests-src
     let bytecode = include_bytes!("./wasm/datastore_manipulations.wasm");
     // create the block containing the erroneous smart contract execution operation
-    let operation = create_execute_sc_operation(&keypair, bytecode, BTreeMap::default()).unwrap();
-    storage.store_operations(vec![operation.clone()]);
-    let block = create_block(
-        KeyPair::generate(0).unwrap(),
+    let operation =
+        ExecutionTestUniverse::create_execute_sc_operation(&keypair, bytecode, BTreeMap::default())
+            .unwrap();
+    universe.storage.store_operations(vec![operation.clone()]);
+    let block = ExecutionTestUniverse::create_block(
+        &keypair,
+        Slot::new(1, 0),
         vec![operation],
         vec![],
-        Slot::new(1, 0),
-    )
-    .unwrap();
+        vec![],
+    );
     // store the block in storage
-    storage.store_block(block.clone());
+    universe.storage.store_block(block.clone());
     // set our block as a final block
     let mut finalized_blocks: HashMap<Slot, BlockId> = Default::default();
     finalized_blocks.insert(block.content.header.content.slot, block.id);
     let block_store = vec![(
         block.id,
         ExecutionBlockMetadata {
-            storage: Some(storage.clone()),
-            same_thread_parent_creator: Some(get_random_address()),
+            storage: Some(universe.storage.clone()),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
         },
     )]
     .into_iter()
     .collect();
-    controller.update_blockclique_status(finalized_blocks, Default::default(), block_store);
+    universe.module_controller.update_blockclique_status(
+        finalized_blocks,
+        Default::default(),
+        block_store,
+    );
     std::thread::sleep(
         exec_cfg
             .t0
@@ -2286,7 +1884,9 @@ fn datastore_manipulations() {
             .into(),
     );
 
-    let events = controller.get_filtered_sc_output_event(EventFilter::default());
+    let events = universe
+        .module_controller
+        .get_filtered_sc_output_event(EventFilter::default());
     // match the events
     println!("{:?}", events);
     assert_eq!(events.len(), 4, "Got {} events, expected 4", events.len());
@@ -2330,7 +1930,8 @@ fn datastore_manipulations() {
     let value_len = ([21, 0, 49].len() + [5, 12, 241].len()) as u64;
 
     assert_eq!(
-        sample_state
+        universe
+            .final_state
             .read()
             .ledger
             .get_balance(&Address::from_public_key(&keypair.get_public_key()))
@@ -2361,54 +1962,27 @@ fn datastore_manipulations() {
                     .saturating_mul_u64(value_len)
             )
     );
-
-    // stop the execution controller
-    manager.stop();
 }
 
 /// This test checks causes a history rewrite in slot sequencing and ensures that emitted events match
 #[test]
 #[serial]
 fn events_from_switching_blockclique() {
-    // Compile the `./wasm_tests` and generate a block with `event_test.wasm`
-    // as data. Then we check if we get an event as expected.
+    // setup the period duration and the maximum gas for asynchronous messages execution
     let exec_cfg = ExecutionConfig {
         t0: MassaTime::from_millis(100),
         cursor_delay: MassaTime::from_millis(0),
         ..ExecutionConfig::default()
     };
-    let storage: Storage = Storage::create_root();
-    let (sample_state, _keep_file, _keep_dir) = get_sample_state(0).unwrap();
-    // init the MIP store
-    let mip_stats_config = MipStatsConfig {
-        block_count_considered: MIP_STORE_STATS_BLOCK_CONSIDERED,
-        warn_announced_version_ratio: Ratio::new_raw(30, 100),
-    };
-    let mip_store = MipStore::try_from(([], mip_stats_config)).unwrap();
+    let block_producer = KeyPair::generate(0).unwrap();
 
-    let slot_execution_output_sender = broadcast::channel(5000).0;
-
-    let channels = ExecutionChannels {
-        slot_execution_output_sender,
-    };
-
-    let (mut manager, controller) = start_execution_worker(
-        exec_cfg.clone(),
-        sample_state.clone(),
-        sample_state.read().pos_state.selector.clone(),
-        mip_store,
-        channels,
-        Arc::new(RwLock::new(create_test_wallet(Some(PreHashMap::default())))),
-        MassaMetrics::new(
-            false,
-            "0.0.0.0:9898".parse().unwrap(),
-            32,
-            std::time::Duration::from_secs(5),
-        )
-        .0,
+    let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
     );
-    // initialize the execution system with genesis blocks
-    init_execution_worker(&exec_cfg, &storage, controller.clone());
+
+    let universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg.clone());
 
     let mut block_metadata: PreHashMap<BlockId, ExecutionBlockMetadata> = Default::default();
     let mut blockclique_blocks: HashMap<Slot, BlockId> = HashMap::new();
@@ -2418,35 +1992,43 @@ fn events_from_switching_blockclique() {
         let blockclique_block_slot = Slot::new(1, 1);
         let keypair = KeyPair::from_str(TEST_SK_2).unwrap();
         let event_test_data = include_bytes!("./wasm/event_test.wasm");
-        let operation =
-            create_execute_sc_operation(&keypair, event_test_data, BTreeMap::default()).unwrap();
-        let blockclique_block = create_block(
-            keypair,
-            vec![operation.clone()],
-            vec![],
-            blockclique_block_slot,
+        let operation = ExecutionTestUniverse::create_execute_sc_operation(
+            &keypair,
+            event_test_data,
+            BTreeMap::default(),
         )
         .unwrap();
+        let blockclique_block = ExecutionTestUniverse::create_block(
+            &block_producer,
+            blockclique_block_slot,
+            vec![operation.clone()],
+            vec![],
+            vec![],
+        );
         blockclique_blocks.insert(blockclique_block_slot, blockclique_block.id);
-        let mut blockclique_block_storage = storage.clone_without_refs();
+        let mut blockclique_block_storage = universe.storage.clone_without_refs();
         blockclique_block_storage.store_block(blockclique_block.clone());
         blockclique_block_storage.store_operations(vec![operation]);
         block_metadata.insert(
             blockclique_block.id,
             ExecutionBlockMetadata {
                 storage: Some(blockclique_block_storage),
-                same_thread_parent_creator: Some(get_random_address()),
+                same_thread_parent_creator: Some(Address::from_public_key(
+                    &keypair.get_public_key(),
+                )),
             },
         );
     }
     // notify execution about blockclique change
-    controller.update_blockclique_status(
+    universe.module_controller.update_blockclique_status(
         Default::default(),
         Some(blockclique_blocks.clone()),
         block_metadata.clone(),
     );
     std::thread::sleep(Duration::from_millis(1000));
-    let events = controller.get_filtered_sc_output_event(EventFilter::default());
+    let events = universe
+        .module_controller
+        .get_filtered_sc_output_event(EventFilter::default());
     assert_eq!(events.len(), 1, "wrong event count");
     assert_eq!(events[0].context.slot, Slot::new(1, 1), "Wrong event slot");
 
@@ -2455,89 +2037,66 @@ fn events_from_switching_blockclique() {
         let blockclique_block_slot = Slot::new(1, 0);
         let keypair = KeyPair::from_str(TEST_SK_1).unwrap();
         let event_test_data = include_bytes!("./wasm/event_test.wasm");
-        let operation =
-            create_execute_sc_operation(&keypair, event_test_data, BTreeMap::default()).unwrap();
-        let blockclique_block = create_block(
-            keypair,
-            vec![operation.clone()],
-            vec![],
-            blockclique_block_slot,
+        let operation = ExecutionTestUniverse::create_execute_sc_operation(
+            &keypair,
+            event_test_data,
+            BTreeMap::default(),
         )
         .unwrap();
+        let blockclique_block = ExecutionTestUniverse::create_block(
+            &keypair,
+            blockclique_block_slot,
+            vec![operation.clone()],
+            vec![],
+            vec![],
+        );
         blockclique_blocks.insert(blockclique_block_slot, blockclique_block.id);
-        let mut blockclique_block_storage = storage.clone_without_refs();
+        let mut blockclique_block_storage = universe.storage.clone_without_refs();
         blockclique_block_storage.store_block(blockclique_block.clone());
         blockclique_block_storage.store_operations(vec![operation]);
         block_metadata.insert(
             blockclique_block.id,
             ExecutionBlockMetadata {
                 storage: Some(blockclique_block_storage),
-                same_thread_parent_creator: Some(get_random_address()),
+                same_thread_parent_creator: Some(Address::from_public_key(
+                    &keypair.get_public_key(),
+                )),
             },
         );
     }
     // notify execution about blockclique change
-    controller.update_blockclique_status(
+    universe.module_controller.update_blockclique_status(
         Default::default(),
         Some(blockclique_blocks.clone()),
         block_metadata.clone(),
     );
     std::thread::sleep(Duration::from_millis(1000));
-    let events = controller.get_filtered_sc_output_event(EventFilter::default());
+    let events = universe
+        .module_controller
+        .get_filtered_sc_output_event(EventFilter::default());
     assert_eq!(events.len(), 2, "wrong event count");
     assert_eq!(events[0].context.slot, Slot::new(1, 0), "Wrong event slot");
     assert_eq!(events[1].context.slot, Slot::new(1, 1), "Wrong event slot");
-
-    manager.stop();
 }
 
 #[test]
 #[serial]
-pub fn not_enough_compilation_gas() {
-    // config
+fn not_enough_compilation_gas() {
+    // setup the period duration
     let exec_cfg = ExecutionConfig {
         t0: MassaTime::from_millis(100),
+        cursor_delay: MassaTime::from_millis(0),
         ..ExecutionConfig::default()
     };
-    // get a sample final state
-    let (sample_state, _keep_file, _keep_dir) = get_sample_state(0).unwrap();
-
-    // init the storage
-    let mut storage = Storage::create_root();
-    // init the MIP store
-    let mip_stats_config = MipStatsConfig {
-        block_count_considered: MIP_STORE_STATS_BLOCK_CONSIDERED,
-        warn_announced_version_ratio: Ratio::new_raw(30, 100),
-    };
-    let mip_store = MipStore::try_from(([], mip_stats_config)).unwrap();
-
-    let slot_execution_output_sender = broadcast::channel(5000).0;
-
-    let channels = ExecutionChannels {
-        slot_execution_output_sender,
-    };
-
-    // start the execution worker
-    let (mut manager, controller) = start_execution_worker(
-        exec_cfg.clone(),
-        sample_state.clone(),
-        sample_state.read().pos_state.selector.clone(),
-        mip_store,
-        channels,
-        Arc::new(RwLock::new(create_test_wallet(Some(PreHashMap::default())))),
-        MassaMetrics::new(
-            false,
-            "0.0.0.0:9898".parse().unwrap(),
-            32,
-            std::time::Duration::from_secs(5),
-        )
-        .0,
-    );
-    // initialize the execution system with genesis blocks
-    init_execution_worker(&exec_cfg, &storage, controller.clone());
-
-    // keypair associated to thread 0
+    let block_producer = KeyPair::generate(0).unwrap();
     let keypair = KeyPair::from_str(TEST_SK_1).unwrap();
+
+    let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
+    );
+    let mut universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg.clone());
 
     // load bytecode
     // you can check the source code of the following wasm file in massa-unit-tests-src
@@ -2558,105 +2117,81 @@ pub fn not_enough_compilation_gas() {
         &keypair,
     )
     .unwrap();
-    storage.store_operations(vec![operation.clone()]);
-    let block = create_block(
-        KeyPair::generate(0).unwrap(),
+    universe.storage.store_operations(vec![operation.clone()]);
+    let block = ExecutionTestUniverse::create_block(
+        &keypair,
+        Slot::new(1, 0),
         vec![operation],
         vec![],
-        Slot::new(1, 0),
-    )
-    .unwrap();
+        vec![],
+    );
     // store the block in storage
-    storage.store_block(block.clone());
+    universe.storage.store_block(block.clone());
     // set our block as a final block
     let mut finalized_blocks: HashMap<Slot, BlockId> = Default::default();
     finalized_blocks.insert(block.content.header.content.slot, block.id);
     let block_store = vec![(
         block.id,
         ExecutionBlockMetadata {
-            storage: Some(storage.clone()),
-            same_thread_parent_creator: Some(get_random_address()),
+            storage: Some(universe.storage.clone()),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
         },
     )]
     .into_iter()
     .collect();
 
     // update blockclique
-    controller.update_blockclique_status(finalized_blocks, Default::default(), block_store);
+    universe.module_controller.update_blockclique_status(
+        finalized_blocks,
+        Default::default(),
+        block_store,
+    );
     std::thread::sleep(Duration::from_millis(100));
 
     // assert events
-    let events = controller.get_filtered_sc_output_event(EventFilter::default());
+    let events = universe
+        .module_controller
+        .get_filtered_sc_output_event(EventFilter::default());
     assert!(events[0]
         .data
         .contains("not enough gas to pay for singlepass compilation"));
-
-    manager.stop();
 }
 
 #[test]
 #[serial]
 fn sc_builtins() {
-    // setup the period duration and the maximum gas for asynchronous messages execution
+    // setup the period duration
     let exec_cfg = ExecutionConfig {
         t0: MassaTime::from_millis(100),
         cursor_delay: MassaTime::from_millis(0),
-        max_async_gas: 100_000,
         ..ExecutionConfig::default()
     };
-    // get a sample final state
-    let (sample_state, _keep_file, _keep_dir) = get_sample_state(0).unwrap();
-
-    // init the MIP store
-    let mip_stats_config = MipStatsConfig {
-        block_count_considered: MIP_STORE_STATS_BLOCK_CONSIDERED,
-        warn_announced_version_ratio: Ratio::new_raw(30, 100),
-    };
-    let mip_store = MipStore::try_from(([], mip_stats_config)).unwrap();
-    // init the storage
-    let mut storage = Storage::create_root();
-
-    let slot_execution_output_sender = broadcast::channel(5000).0;
-
-    let channels = ExecutionChannels {
-        slot_execution_output_sender,
-    };
-
-    // start the execution worker
-    let (mut manager, controller) = start_execution_worker(
-        exec_cfg.clone(),
-        sample_state.clone(),
-        sample_state.read().pos_state.selector.clone(),
-        mip_store,
-        channels,
-        Arc::new(RwLock::new(create_test_wallet(Some(PreHashMap::default())))),
-        MassaMetrics::new(
-            false,
-            "0.0.0.0:9898".parse().unwrap(),
-            32,
-            std::time::Duration::from_secs(5),
-        )
-        .0,
-    );
-    // initialize the execution system with genesis blocks
-    init_execution_worker(&exec_cfg, &storage, controller.clone());
-    // keypair associated to thread 0
+    let block_producer = KeyPair::generate(0).unwrap();
     let keypair = KeyPair::from_str(TEST_SK_1).unwrap();
+
+    let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
+    );
+    let mut universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg.clone());
     // load bytecode
     // you can check the source code of the following wasm file in massa-unit-tests-src
     let bytecode = include_bytes!("./wasm/use_builtins.wasm");
     // create the block containing the erroneous smart contract execution operation
-    let operation = create_execute_sc_operation(&keypair, bytecode, BTreeMap::default()).unwrap();
-    storage.store_operations(vec![operation.clone()]);
-    let block = create_block(
-        KeyPair::generate(0).unwrap(),
+    let operation =
+        ExecutionTestUniverse::create_execute_sc_operation(&keypair, bytecode, BTreeMap::default())
+            .unwrap();
+    universe.storage.store_operations(vec![operation.clone()]);
+    let block = ExecutionTestUniverse::create_block(
+        &keypair,
+        Slot::new(1, 0),
         vec![operation],
         vec![],
-        Slot::new(1, 0),
-    )
-    .unwrap();
+        vec![],
+    );
     // store the block in storage
-    storage.store_block(block.clone());
+    universe.storage.store_block(block.clone());
     // set our block as a final block
     let mut finalized_blocks: HashMap<Slot, BlockId> = Default::default();
     finalized_blocks.insert(block.content.header.content.slot, block.id);
@@ -2664,11 +2199,11 @@ fn sc_builtins() {
     block_metadata.insert(
         block.id,
         ExecutionBlockMetadata {
-            same_thread_parent_creator: Some(get_random_address()),
-            storage: Some(storage),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
+            storage: Some(universe.storage.clone()),
         },
     );
-    controller.update_blockclique_status(
+    universe.module_controller.update_blockclique_status(
         finalized_blocks,
         Default::default(),
         block_metadata.clone(),
@@ -2676,7 +2211,9 @@ fn sc_builtins() {
     std::thread::sleep(Duration::from_millis(10));
 
     // retrieve the event emitted by the execution error
-    let events = controller.get_filtered_sc_output_event(EventFilter::default());
+    let events = universe
+        .module_controller
+        .get_filtered_sc_output_event(EventFilter::default());
     // match the events
     assert!(!events.is_empty(), "One event was expected");
     assert!(events[0].data.contains("massa_execution_error"));
@@ -2689,85 +2226,53 @@ fn sc_builtins() {
         .contains("abort with date and rnd at use_builtins.ts:0 col: 0"));
 
     assert_eq!(
-        sample_state
+        universe
+            .final_state
             .read()
             .ledger
             .get_balance(&Address::from_public_key(&keypair.get_public_key()))
             .unwrap(),
         Amount::from_str("299990").unwrap()
     );
-    // stop the execution controller
-    manager.stop();
 }
 
 #[test]
 #[serial]
 fn validate_address() {
-    // setup the period duration and the maximum gas for asynchronous messages execution
+    // setup the period duration
     let exec_cfg = ExecutionConfig {
         t0: MassaTime::from_millis(100),
         cursor_delay: MassaTime::from_millis(0),
-        max_async_gas: 100_000,
         ..ExecutionConfig::default()
     };
-    // get a sample final state
-    let (sample_state, _keep_file, _keep_dir) = get_sample_state(0).unwrap();
-
-    // init the MIP store
-    let mip_stats_config = MipStatsConfig {
-        block_count_considered: MIP_STORE_STATS_BLOCK_CONSIDERED,
-        warn_announced_version_ratio: Ratio::new_raw(30, 100),
-    };
-    let mip_store = MipStore::try_from(([], mip_stats_config)).unwrap();
-    // init the storage
-    let mut storage = Storage::create_root();
-
-    let slot_execution_output_sender = broadcast::channel(5000).0;
-
-    let channels = ExecutionChannels {
-        slot_execution_output_sender,
-    };
-
-    // start the execution worker
-    let (mut manager, controller) = start_execution_worker(
-        exec_cfg.clone(),
-        sample_state.clone(),
-        sample_state.read().pos_state.selector.clone(),
-        mip_store,
-        channels,
-        Arc::new(RwLock::new(create_test_wallet(Some(PreHashMap::default())))),
-        MassaMetrics::new(
-            false,
-            "0.0.0.0:9898".parse().unwrap(),
-            32,
-            std::time::Duration::from_secs(5),
-        )
-        .0,
-    );
-    // initialize the execution system with genesis blocks
-    init_execution_worker(&exec_cfg, &storage, controller.clone());
-
-    // keypair associated to thread 0
+    let block_producer = KeyPair::generate(0).unwrap();
     let keypair = KeyPair::from_str(TEST_SK_1).unwrap();
-    // let address = Address::from_public_key(&keypair.get_public_key());
+
+    let mut foreign_controllers = ExecutionForeignControllers::new_with_mocks();
+    selector_boilerplate(
+        &mut foreign_controllers.selector_controller,
+        &block_producer,
+    );
+    let mut universe = ExecutionTestUniverse::new(foreign_controllers, exec_cfg.clone());
 
     // load bytecode
     // you can check the source code of the following wasm file in massa-unit-tests-src
     let bytecode = include_bytes!("./wasm/validate_address.wasm");
 
     // create the block containing the erroneous smart contract execution operation
-    let operation = create_execute_sc_operation(&keypair, bytecode, BTreeMap::default()).unwrap();
-    storage.store_operations(vec![operation.clone()]);
-    let block = create_block(
-        KeyPair::generate(0).unwrap(),
+    let operation =
+        ExecutionTestUniverse::create_execute_sc_operation(&keypair, bytecode, BTreeMap::default())
+            .unwrap();
+    universe.storage.store_operations(vec![operation.clone()]);
+    let block = ExecutionTestUniverse::create_block(
+        &keypair,
+        Slot::new(1, 0),
         vec![operation],
         vec![],
-        Slot::new(1, 0),
-    )
-    .unwrap();
-
+        vec![],
+    );
     // store the block in storage
-    storage.store_block(block.clone());
+    universe.storage.store_block(block.clone());
 
     // set our block as a final block
     let mut finalized_blocks: HashMap<Slot, BlockId> = Default::default();
@@ -2775,13 +2280,17 @@ fn validate_address() {
     let block_store = vec![(
         block.id,
         ExecutionBlockMetadata {
-            storage: Some(storage.clone()),
-            same_thread_parent_creator: Some(get_random_address()),
+            storage: Some(universe.storage.clone()),
+            same_thread_parent_creator: Some(Address::from_public_key(&keypair.get_public_key())),
         },
     )]
     .into_iter()
     .collect();
-    controller.update_blockclique_status(finalized_blocks, Default::default(), block_store);
+    universe.module_controller.update_blockclique_status(
+        finalized_blocks,
+        Default::default(),
+        block_store,
+    );
     std::thread::sleep(
         exec_cfg
             .t0
@@ -2789,7 +2298,9 @@ fn validate_address() {
             .into(),
     );
 
-    let events = controller.get_filtered_sc_output_event(EventFilter::default());
+    let events = universe
+        .module_controller
+        .get_filtered_sc_output_event(EventFilter::default());
     // match the events
     assert_eq!(events.len(), 2);
     assert!(
@@ -2804,7 +2315,8 @@ fn validate_address() {
     );
 
     assert_eq!(
-        sample_state
+        universe
+            .final_state
             .read()
             .ledger
             .get_balance(&Address::from_public_key(&keypair.get_public_key()))
@@ -2814,64 +2326,4 @@ fn validate_address() {
             // Gas fee
             .saturating_sub(Amount::const_init(10, 0))
     );
-
-    // stop the execution controller
-    manager.stop();
-}
-
-#[test]
-fn test_take_batch() {
-    let final_state = get_sample_state(0).unwrap().0;
-    let active_history = Arc::new(RwLock::new(ActiveHistory::default()));
-
-    let mut speculative_pool = SpeculativeAsyncPool::new(final_state, active_history);
-
-    let address = Address::User(UserAddress::UserAddressV0(UserAddressV0(
-        Hash::compute_from(b"abc"),
-    )));
-
-    for i in 1..10 {
-        let message = AsyncMessage::new(
-            Slot::new(0, 0),
-            0,
-            address,
-            address,
-            "function".to_string(),
-            i,
-            Amount::from_str("0.1").unwrap(),
-            Amount::from_str("0.3").unwrap(),
-            Slot::new(1, 0),
-            Slot::new(3, 0),
-            Vec::new(),
-            None,
-            None,
-        );
-        speculative_pool.push_new_message(message)
-    }
-    assert_eq!(speculative_pool.get_message_infos().len(), 9);
-    speculative_pool.take_batch_to_execute(Slot::new(2, 0), 19);
-    assert_eq!(speculative_pool.get_message_infos().len(), 4);
-}
-
-fn create_execute_sc_operation(
-    sender_keypair: &KeyPair,
-    data: &[u8],
-    datastore: Datastore,
-) -> Result<SecureShareOperation, ExecutionError> {
-    let op = OperationType::ExecuteSC {
-        data: data.to_vec(),
-        max_gas: 100_000_000,
-        max_coins: Amount::from_str("5000000").unwrap(),
-        datastore,
-    };
-    let op = Operation::new_verifiable(
-        Operation {
-            fee: Amount::const_init(10, 0),
-            expire_period: 10,
-            op,
-        },
-        OperationSerializer::new(),
-        sender_keypair,
-    )?;
-    Ok(op)
 }
